%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaMicroarrayDataSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.AromaMicroarrayDataSet.AromaMicroarrayDataSet}
\alias{as.AromaMicroarrayDataSet.AromaMicroarrayDataSet}
\alias{AromaMicroarrayDataSet.as.AromaMicroarrayDataSet}
\alias{as.AromaMicroarrayDataSet.AromaMicroarrayDataSet}
\alias{as.AromaMicroarrayDataSet,AromaMicroarrayDataSet-method}

\alias{as.AromaMicroarrayDataSet.list}
\alias{as.AromaMicroarrayDataSet.default}

\title{Coerce an object to an AromaMicroarrayDataSet object}

\description{
  Coerce an object to an AromaMicroarrayDataSet object.
}

\usage{\method{as.AromaMicroarrayDataSet}{AromaMicroarrayDataSet}(object, ...)}

\arguments{
 \item{...}{Other arguments passed to \code{\link[base]{list.files}}().}
}

\value{
  Returns an \code{\link{AromaMicroarrayDataSet}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AromaMicroarrayDataSet}}.
}
\keyword{internal}
\keyword{methods}
