%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Explorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{display.Explorer}
\alias{display.Explorer}
\alias{Explorer.display}
\alias{display,Explorer-method}

\title{Displays the explorer in the default browser}

\description{
 Displays the explorer in the default browser.
}

\usage{
\method{display}{Explorer}(this, filename=sprintf("\%s.html", class(this)[1]), ..., verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Explorer}}.
}
\keyword{internal}
\keyword{methods}
