%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  findAnnotationDataByChipType.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{findAnnotationDataByChipType}
\alias{findAnnotationDataByChipType.default}
\alias{findAnnotationDataByChipType}

 \title{Locates an annotation data file by its chip type}

 \description{
  Locates an annotation data file by its chip type.
 }

 \usage{
\method{findAnnotationDataByChipType}{default}(chipType, pattern=chipType, ...)
}

 \arguments{
   \item{chipType}{A \code{\link[base]{character}} string.}
   \item{pattern}{A filename pattern to search for.}
   \item{...}{Arguments passed to \code{\link{findAnnotationData}}().}
 }

 \author{Henrik Bengtsson}


\keyword{internal}
