%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Explorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAlias.Explorer}
\alias{getAlias.Explorer}
\alias{Explorer.getAlias}
\alias{getAlias,Explorer-method}

\title{Gets the alias of the output set}

\description{
  Gets the alias of the output set.
}

\usage{
\method{getAlias}{Explorer}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}, or \code{\link[base]{NULL}} if no alias is set.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[aroma.core:setAlias.Explorer]{*setAlias}()}.
  \code{\link[aroma.core:getName.Explorer]{*getName}()}.
  For more information see \code{\link{Explorer}}.
}
\keyword{internal}
\keyword{methods}
