%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaPlatformInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAromaPlatform.AromaPlatformInterface}
\alias{getAromaPlatform.AromaPlatformInterface}
\alias{AromaPlatformInterface.getAromaPlatform}
\alias{getAromaPlatform,AromaPlatformInterface-method}

\title{Gets the platform}

\description{
 Gets the platform object.
}

\usage{
\method{getAromaPlatform}{AromaPlatformInterface}(this, ..., force=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{force}{If \code{\link[base:logical]{TRUE}}, any cached results are ignored, otherwise not.}
}

\value{
 Returns an \code{\link{AromaPlatform}} object.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[aroma.core:getPlatform.AromaPlatformInterface]{*getPlatform}()}.
  For more information see \code{\link{AromaPlatformInterface}}.
}
\keyword{internal}
\keyword{methods}
