%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Explorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getArrays.Explorer}
\alias{getArrays.Explorer}
\alias{Explorer.getArrays}
\alias{getArrays.Explorer}
\alias{getArrays,Explorer-method}


\title{Gets the names of the arrays}

\description{
 Gets the names of the arrays in the explorer.
}

\usage{\method{getArrays}{Explorer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Explorer}}.
}
\keyword{internal}
\keyword{methods}
