%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaMicroarrayDataSetTuple.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getArrayTuple.AromaMicroarrayDataSetTuple}
\alias{getArrayTuple.AromaMicroarrayDataSetTuple}
\alias{AromaMicroarrayDataSetTuple.getArrayTuple}
\alias{getArrayTuple.AromaMicroarrayDataSetTuple}
\alias{getArrayTuple,AromaMicroarrayDataSetTuple-method}


\title{Gets arrays across chip types for one sample}

\description{
 Gets arrays across chip types for one sample.
}

\usage{\method{getArrayTuple}{AromaMicroarrayDataSetTuple}(this, array, ...)}

\arguments{
  \item{array}{Sample of interest.}
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}} of \code{\link{AromaMicroarrayDataSet}}:s.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AromaMicroarrayDataSetTuple}}.
}
\keyword{internal}
\keyword{methods}
