%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaPlatformInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipType.AromaPlatformInterface}
\alias{getChipType.AromaPlatformInterface}
\alias{AromaPlatformInterface.getChipType}
\alias{getChipType,AromaPlatformInterface-method}

\title{Gets the chip type}

\description{
 Gets the chip type.
}

\usage{
\method{getChipType}{AromaPlatformInterface}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AromaPlatformInterface}}.
}
\keyword{internal}
\keyword{methods}
