%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomalModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChromosomes.ChromosomalModel}
\alias{getChromosomes.ChromosomalModel}
\alias{ChromosomalModel.getChromosomes}
\alias{getChromosomes,ChromosomalModel-method}
\alias{setChromosomes.ChromosomalModel}

\title{Gets the chromosomes to be processed}

\description{
 Gets the chromosomes to be processed.
}

\usage{
\method{getChromosomes}{ChromosomalModel}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{ChromosomalModel}}.
}
\keyword{internal}
\keyword{methods}
