%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaTransform.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInputDataSet.AromaTransform}
\alias{getInputDataSet.AromaTransform}
\alias{AromaTransform.getInputDataSet}
\alias{getInputDataSet,AromaTransform-method}

\title{Gets the input data set}

\description{
 Gets the input data set that is to be (or has been) transformed.
}

\usage{
\method{getInputDataSet}{AromaTransform}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AromaMicroarrayDataSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AromaTransform}}.
}
\keyword{internal}
\keyword{methods}
