%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomalModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getNames.ChromosomalModel}
\alias{getNames.ChromosomalModel}
\alias{ChromosomalModel.getNames}
\alias{getNames.ChromosomalModel}
\alias{getNames,ChromosomalModel-method}


\title{Gets the names of the arrays}

\description{
 Gets the names of the arrays available to the model.
}

\usage{\method{getNames}{ChromosomalModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomalModel}}.
}
\keyword{internal}
\keyword{methods}
