%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParametersInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getParameters.ParametersInterface}
\alias{getParameters.ParametersInterface}
\alias{ParametersInterface.getParameters}
\alias{getParameters,ParametersInterface-method}

\title{Gets a list of parameters}

\description{
 Gets a list of parameters associated with the object.
}

\usage{
\method{getParameters}{ParametersInterface}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{ParametersInterface}}.
}
\keyword{internal}
\keyword{methods}
