%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaPlatformInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPlatform.AromaPlatformInterface}
\alias{getPlatform.AromaPlatformInterface}
\alias{AromaPlatformInterface.getPlatform}
\alias{getPlatform,AromaPlatformInterface-method}

\title{Gets the platform}

\description{
 Gets the platform label.
}

\usage{
\method{getPlatform}{AromaPlatformInterface}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[aroma.core:getAromaPlatform.AromaPlatformInterface]{*getAromaPlatform}()}.
  For more information see \code{\link{AromaPlatformInterface}}.
}
\keyword{internal}
\keyword{methods}
