%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaTransform.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRootPath.AromaTransform}
\alias{getRootPath.AromaTransform}
\alias{AromaTransform.getRootPath}
\alias{getRootPath.AromaTransform}
\alias{getRootPath,AromaTransform-method}


\title{Gets the root path of the output directory}

\description{
 Gets the root path of the output directory that is returned by \code{\link[aroma.core:getPath.AromaTransform]{*getPath}()}.
 A root path is a directory in the current working directory.
}

\usage{\method{getRootPath}{AromaTransform}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.core:getPath.AromaTransform]{*getPath}()}.
  For more information see \code{\link{AromaTransform}}.
}
\keyword{internal}
\keyword{methods}
