%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Explorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRootPath.Explorer}
\alias{getRootPath.Explorer}
\alias{Explorer.getRootPath}
\alias{getRootPath,Explorer-method}

\title{Gets the root path of the output directory}

\description{
 Gets the root path of the output directory that is returned by \code{\link[aroma.core:getPath.Explorer]{*getPath}()}.
 A root path is a directory in the current working directory.
}

\usage{
\method{getRootPath}{Explorer}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.core:getPath.Explorer]{*getPath}()}.
  For more information see \code{\link{Explorer}}.
}
\keyword{internal}
\keyword{methods}
