%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaTransform.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTags.AromaTransform}
\alias{getTags.AromaTransform}
\alias{AromaTransform.getTags}
\alias{getTags,AromaTransform-method}

\title{Gets the tags of the output data set}

\description{
 Gets the tags of the output data set, which equals the tags of the input data set plus the tags
 of this transformation.
}

\usage{
\method{getTags}{AromaTransform}(this, collapse=NULL, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AromaTransform}}.
}
\keyword{internal}
\keyword{methods}
