%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Interface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Interface}
\docType{class}
\alias{Interface}


 \title{The Interface class}

 \description{
  Package:  aroma.core \cr
\bold{Class Interface}\cr

\code{+--}\code{Interface}\cr

\bold{Directly known subclasses:}\cr
\emph{\link[aroma.core]{AromaPlatformInterface}}, \emph{\link[aroma.core]{UnitNamesFile}}\cr

public class \bold{Interface}\cr



  This class represents a special set of classes whose purpose is to
  provide methods (but not fields) shared by multiple different classes.
 }

 \usage{Interface(core=NA, ...)}

 \arguments{
   \item{core}{The core value.}
   \item{...}{Not used.}
 }

 \section{Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{extend} \tab  -\cr
 \tab \code{print} \tab  -\cr
 \tab \code{uses} \tab  -\cr
}



 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{classes}
\keyword{internal}
