%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaPlatformInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isCompatibleWith.AromaPlatformInterface}
\alias{isCompatibleWith.AromaPlatformInterface}
\alias{AromaPlatformInterface.isCompatibleWith}
\alias{isCompatibleWith,AromaPlatformInterface-method}

\title{Checks if a particular unit annotation data file is compatible}

\description{
 Checks if a particular unit annotation data file is compatible with this AromaPlatformInterface class.
}

\usage{
\method{isCompatibleWith}{AromaPlatformInterface}(this, udf, ...)
}

\arguments{
  \item{udf}{An unit annotation data file.}
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if compatible and \code{\link[base:logical]{FALSE}} otherwise.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[aroma.core:getPlatform.AromaPlatformInterface]{*getPlatform}()}.
  For more information see \code{\link{AromaPlatformInterface}}.
}
\keyword{internal}
\keyword{methods}
