%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Explorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfArrays.Explorer}
\alias{nbrOfArrays.Explorer}
\alias{Explorer.nbrOfArrays}
\alias{nbrOfArrays,Explorer-method}

\title{Gets the total number of arrays}

\description{
 Gets the total number of arrays considered by the explorer.
}

\usage{
\method{nbrOfArrays}{Explorer}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Explorer}}.
}
\keyword{internal}
\keyword{methods}
