%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{[.AromaTabularBinaryFile}
\alias{[[.AromaTabularBinaryFile}
\alias{[<-.AromaTabularBinaryFile}
\alias{allocate}
\alias{append}
\alias{append.default}
\alias{append.GenericDataFileSet}
\alias{appendFiles}
\alias{appendFiles.GenericDataFileSet}
\alias{apply}
\alias{apply.default}
\alias{apply.SampleAnnotationFile}
\alias{as.character.AromaTabularBinaryFile}
\alias{as.character.GenericTabularFile}
\alias{calcMargins}
\alias{calcMargins.matrix}
\alias{clearCache.AromaTabularBinaryFile}
\alias{clearCache.GenericDataFileSet}
\alias{clone.GenericDataFile}
\alias{clone.GenericDataFileSet}
\alias{colMeans}
\alias{colMeans.AromaTabularBinaryFile}
\alias{colMeans.default}
\alias{colMedians.AromaTabularBinaryFile}
\alias{colStats}
\alias{colStats.AromaTabularBinaryFile}
\alias{colSums}
\alias{colSums.AromaTabularBinaryFile}
\alias{colSums.default}
\alias{compareChecksum}
\alias{compareChecksum.GenericDataFile}
\alias{convertTable}
\alias{convertTable.default}
\alias{copyTo}
\alias{copyTo.GenericDataFile}
\alias{copyTo.GenericDataFileSet}
\alias{digest2}
\alias{dim.AromaTabularBinaryFile}
\alias{dim.GenericTabularFile}
\alias{dimnames<-.AromaTabularBinaryFile}
\alias{equals.GenericDataFile}
\alias{findFiles}
\alias{findPngDevice}
\alias{findPngDevice.default}
\alias{findSAFs}
\alias{findSAFs.SampleAnnotationSet}
\alias{fit2d}
\alias{fit2d.matrix}
\alias{fromFile}
\alias{fromFile.GenericDataFile}
\alias{fromFiles}
\alias{fromPath}
\alias{fromPath.SampleAnnotationFile}
\alias{fromPath.SampleAnnotationSet}
\alias{GenericSummary}
\alias{GenericTabularFile}
\alias{getAlias}
\alias{getAlias.GenericDataFile}
\alias{getAttribute}
\alias{getAttribute.GenericDataFile}
\alias{getAttributes}
\alias{getAttributes.GenericDataFile}
\alias{getBytesPerColumn}
\alias{getBytesPerColumn.AromaTabularBinaryFile}
\alias{getChecksum}
\alias{getChecksum.GenericDataFile}
\alias{getChromosome}
\alias{getChromosome.Arguments}
\alias{getChromosomes}
\alias{getChromosomes.Arguments}
\alias{getColClasses}
\alias{getColClasses.AromaTabularBinaryFile}
\alias{getColumnNames}
\alias{getColumnNames.AromaTabularBinaryFile}
\alias{getColumnNames.GenericTabularFile}
\alias{getColumnNameTranslator}
\alias{getColumnNameTranslator.GenericTabularFile}
\alias{getCommonListElements}
\alias{getCommonPrefix}
\alias{getDeviceResolution}
\alias{getDeviceResolution.default}
\alias{getFile}
\alias{getFile.GenericDataFileSet}
\alias{getFilename}
\alias{getFilename.Arguments}
\alias{getFiles}
\alias{getFiles.GenericDataFileSet}
\alias{getFileSize}
\alias{getFileSize.GenericDataFile}
\alias{getFileSize.GenericDataFileSet}
\alias{getFileType}
\alias{getFullName}
\alias{getFullNames}
\alias{getFullNames.GenericDataFileSet}
\alias{getHeader.GenericTabularFile}
\alias{getLabel.GenericDataFile}
\alias{getNames}
\alias{getPathname}
\alias{getPathnames}
\alias{getPatterns}
\alias{getPatterns.SampleAnnotationFile}
\alias{getReadArguments}
\alias{getReadArguments.GenericTabularFile}
\alias{getTags}
\alias{hasTag}
\alias{hasTag.GenericDataFile}
\alias{hasTag.GenericDataFileSet}
\alias{hasTags}
\alias{hasTags.GenericDataFile}
\alias{hasTags.GenericDataFileSet}
\alias{importFromTable}
\alias{importFromTable.FileMatrix}
\alias{indexOf}
\alias{indexOf.GenericDataFileSet}
\alias{isFile.GenericDataFile}
\alias{lapply.AromaTabularBinaryFile}
\alias{lapplyInChunks}
\alias{lapplyInChunks.numeric}
\alias{library}
\alias{listToXml}
\alias{listToXml.list}
\alias{locallyUnique}
\alias{locallyUnique.default}
\alias{matchPatterns}
\alias{matchPatterns.SampleAnnotationFile}
\alias{matrixBlockPolish}
\alias{mergeByCommonTails}
\alias{nbrOfColumns}
\alias{nbrOfColumns.AromaTabularBinaryFile}
\alias{nbrOfColumns.GenericTabularFile}
\alias{nbrOfFiles}
\alias{nbrOfLines}
\alias{nbrOfLines.GenericTabularFile}
\alias{nbrOfRows}
\alias{nbrOfRows.AromaTabularBinaryFile}
\alias{nbrOfRows.GenericTabularFile}
\alias{norm2d}
\alias{norm2d.matrix}
\alias{print.GenericSummary}
\alias{readChecksum}
\alias{readChecksum.GenericDataFile}
\alias{readColumnNames}
\alias{readColumnNames.GenericTabularFile}
\alias{readData}
\alias{readData.AromaTabularBinaryFile}
\alias{readData.GenericTabularFile}
\alias{readData.SampleAnnotationFile}
\alias{readDataFrame}
\alias{readDataFrame.AromaTabularBinaryFile}
\alias{readDataFrame.GenericTabularFile}
\alias{readDataFrame.SampleAnnotationFile}
\alias{readFooter}
\alias{readHeader}
\alias{readHeader.AromaTabularBinaryFile}
\alias{readHeader.GenericTabularFile}
\alias{readLines}
\alias{readLines.default}
\alias{readLines.GenericTabularFile}
\alias{readRawFooter}
\alias{readRawFooter.AromaTabularBinaryFile}
\alias{renameTo}
\alias{renameTo.GenericDataFile}
\alias{renameToUpperCaseExt}
\alias{renameToUpperCaseExt.GenericDataFile}
\alias{reorder.GenericDataFileSet}
\alias{require}
\alias{SampleAnnotationFile}
\alias{SampleAnnotationSet}
\alias{sapply}
\alias{sapply.default}
\alias{sapply.GenericDataFileSet}
\alias{setAlias}
\alias{setAlias.GenericDataFile}
\alias{setAttribute}
\alias{setAttribute.GenericDataFile}
\alias{setAttributes}
\alias{setAttributes.GenericDataFile}
\alias{setAttributesBy}
\alias{setAttributesBy.GenericDataFile}
\alias{setAttributesByTags}
\alias{setAttributesByTags.AromaTabularBinaryFile}
\alias{setAttributesByTags.GenericDataFile}
\alias{setColumnNameTranslator}
\alias{setColumnNameTranslator.GenericTabularFile}
\alias{setLabel.GenericDataFile}
\alias{setTags}
\alias{splitByCommonTails}
\alias{splitInChunks}
\alias{splitInChunks.numeric}
\alias{stext}
\alias{stext.default}
\alias{stringTree}
\alias{stringTree.character}
\alias{subset.AromaTabularBinaryFile}
\alias{summary.AromaTabularBinaryFile}
\alias{testAttributes}
\alias{testAttributes.GenericDataFile}
\alias{translateColumnNames}
\alias{translateColumnNames.GenericTabularFile}
\alias{updateData}
\alias{updateData.AromaTabularBinaryFile}
\alias{updateDataColumn}
\alias{updateDataColumn.AromaTabularBinaryFile}
\alias{validateChecksum}
\alias{validateChecksum.GenericDataFile}
\alias{verify}
\alias{verify.GenericTabularFile}
\alias{writeChecksum}
\alias{writeChecksum.GenericDataFile}
\alias{writeColumnsToFiles}
\alias{writeColumnsToFiles.GenericTabularFile}
\alias{writeFooter}
\alias{writeRawFooter}
\alias{writeRawFooter.AromaTabularBinaryFile}
\alias{xmlToList}
\alias{xmlToList.character}
\alias{yellow.colors}
\alias{as.character.Interface}
\alias{as.GrayscaleImage}
\alias{as.TrueColorImage}
\alias{as.TrueColorImage.Image}
\alias{as.TrueColorImage.matrix}
\alias{backtransformGenotypeCone}
\alias{backtransformGenotypeCone.matrix}
\alias{colorize}
\alias{colorize.Image}
\alias{display}
\alias{display.Image}
\alias{extend.Interface}
\alias{fitGenotypeCone}
\alias{fitSplineBlockPolish}
\alias{fitSplineBlockPolish.matrix}
\alias{fitWHLAPLM}
\alias{fitWHLAPLM.matrix}
\alias{fitWHRCModel}
\alias{fitWHRCModel.matrix}
\alias{fitWLAPLM}
\alias{fitWLAPLM.matrix}
\alias{fitWRCModel}
\alias{fitWRCModel.matrix}
\alias{fitWRMA}
\alias{fitWRMA.matrix}
\alias{gaussianSmoothing}
\alias{gaussianSmoothing.matrix}
\alias{gaussianSmoothing.numeric}
\alias{gaussKernel}
\alias{getImage}
\alias{getImage.matrix}
\alias{interleave}
\alias{interleave.Image}
\alias{log2abs}
\alias{log2center}
\alias{log2neg}
\alias{log2pos}
\alias{mergeBoxplotStats}
\alias{normalizeGenotypeCone}
\alias{normalizeGenotypeCone.matrix}
\alias{print.Interface}
\alias{rescale}
\alias{rescale.Image}
\alias{rgbTransform}
\alias{rgbTransform.Image}
\alias{smoothWRMA}
\alias{smoothWRMA.matrix}
\alias{smoothWSA}
\alias{smoothWSA.matrix}
\alias{sqrtabs}
\alias{sqrtcenter}
\alias{sqrtneg}
\alias{sqrtpos}
\alias{uses}
\alias{uses.Interface}
\alias{uses.character}
\alias{whatDataType}
\alias{whatDataType.default}
\alias{clearCache.AromaMicroarrayDataSet}
\alias{getChipType}
\alias{getChipType.AromaMicroarrayDataFile}
\alias{getChipType.AromaMicroarrayDataSet}
\alias{getLabel.AromaMicroarrayDataFile}
\alias{getPlatform}
\alias{getPlatform.AromaMicroarrayDataFile}
\alias{getPlatform.AromaMicroarrayDataSet}
\alias{setLabel.AromaMicroarrayDataFile}
\alias{AromaMicroarrayDataFileSet}
\alias{byName}
\alias{byName.GenericDataFileSet}
\alias{extractMatrix}
\alias{extractMatrix.AromaTabularBinaryFile}
\alias{extractMatrix.AromaTabularBinarySet}
\alias{findByName}
\alias{findByName.GenericDataFileSet}
\alias{getFileClass}
\alias{getFileClass.GenericDataFileSet}
\alias{validate.AromaMicroarrayDataSet}
\alias{validate.GenericDataFileSet}
\alias{applyRegions}
\alias{applyRegions.CopyNumberOutliers}
\alias{applyRows}
\alias{applyRows.CopyNumberRegions}
\alias{as.character.AromaMicroarrayDataSetTuple}
\alias{as.character.CopyNumberOutliers}
\alias{as.character.CopyNumberRegions}
\alias{as.character.RawCopyNumbers}
\alias{as.character.TabularTextFile}
\alias{as.data.frame.CopyNumberOutliers}
\alias{as.data.frame.CopyNumberRegions}
\alias{as.data.frame.RawCopyNumbers}
\alias{asMatrixOfFiles}
\alias{asMatrixOfFiles.AromaMicroarrayDataSetTuple}
\alias{byPath}
\alias{byPath.AromaMicroarrayDataSetTuple}
\alias{CopyNumberOutliers}
\alias{CopyNumberRegions}
\alias{drawLevels}
\alias{drawLevels.CopyNumberOutliers}
\alias{drawLevels.CopyNumberRegions}
\alias{equals.GenericDataFileSet}
\alias{estimateStandardDeviation}
\alias{extractCNRs}
\alias{extractCNRs.default}
\alias{extractMatrix.GenericTabularFile}
\alias{extractMatrix.GenericTabularFileSet}
\alias{extractRawCNs}
\alias{extractRawCNs.default}
\alias{gaussianSmoothing.RawCopyNumbers}
\alias{getAlias.AromaMicroarrayDataSetTuple}
\alias{getArrays}
\alias{getArrayTuple}
\alias{getAsteriskTags}
\alias{getAsteriskTags.AromaMicroarrayDataSetTuple}
\alias{getChipTypes}
\alias{getCNs}
\alias{getCNs.RawCopyNumbers}
\alias{getColumnNames.TabularTextFile}
\alias{getFullName.AromaMicroarrayDataSetTuple}
\alias{getFullNames.AromaMicroarrayDataSetTuple}
\alias{getFullNameTranslator}
\alias{getFullNameTranslator.GenericDataFile}
\alias{getHeader.TabularTextFile}
\alias{getListOfCdfs}
\alias{getListOfCdfs.AromaMicroarrayDataSetTuple}
\alias{getListOfSets}
\alias{getListOfSets.AromaMicroarrayDataSetTuple}
\alias{getLociFields}
\alias{getLociFields.RawCopyNumbers}
\alias{getName.AromaMicroarrayDataSetTuple}
\alias{getNames.AromaMicroarrayDataSetTuple}
\alias{getPhysicalPositions}
\alias{getPhysicalPositions.RawCopyNumbers}
\alias{getReadArguments.TabularTextFile}
\alias{getTableOfArrays}
\alias{getTags.AromaMicroarrayDataSetTuple}
\alias{getXY}
\alias{getXY.RawCopyNumbers}
\alias{gunzip.GenericDataFile}
\alias{gzip.GenericDataFile}
\alias{hasFile}
\alias{hasFile.GenericDataFileSet}
\alias{indexOfArrays}
\alias{lines.CopyNumberOutliers}
\alias{lines.CopyNumberRegions}
\alias{lines.RawCopyNumbers}
\alias{nbrOfArrays}
\alias{nbrOfChipTypes}
\alias{nbrOfLines.TabularTextFile}
\alias{nbrOfLoci}
\alias{nbrOfLoci.RawCopyNumbers}
\alias{nbrOfRegions}
\alias{nbrOfRegions.CopyNumberOutliers}
\alias{nbrOfRegions.CopyNumberRegions}
\alias{nbrOfRows.TabularTextFile}
\alias{plot.RawCopyNumbers}
\alias{points.RawCopyNumbers}
\alias{readColumnNames.TabularTextFile}
\alias{readColumns}
\alias{readColumns.AromaTabularBinaryFile}
\alias{readColumns.GenericTabularFile}
\alias{readColumns.TabularTextFile}
\alias{readDataFrame.TabularTextFile}
\alias{readHeader.TabularTextFile}
\alias{readLines.TabularTextFile}
\alias{setAlias.AromaMicroarrayDataSetTuple}
\alias{setFullNameTranslator}
\alias{setFullNameTranslator.GenericDataFile}
\alias{setTags.AromaMicroarrayDataSetTuple}
\alias{sort.RawCopyNumbers}
\alias{stextChipType}
\alias{stextChipType.character}
\alias{summary.RawCopyNumbers}
\alias{TabularTextFile}
\alias{translateFullName}
\alias{translateFullName.GenericDataFile}
\alias{verify.TabularTextFile}
\alias{writeColumnsToFiles.TabularTextFile}
\alias{extractCopyNumberRegions}
\alias{extractCopyNumberRegions.default}
\alias{extractRawCopyNumbers}
\alias{extractRawCopyNumbers.default}
\alias{getAttributeXY}
\alias{getAttributeXY.AromaMicroarrayDataFile}
\alias{getPloidy}
\alias{getPloidy.AromaMicroarrayDataFile}
\alias{hasAttributeXY}
\alias{hasAttributeXY.AromaMicroarrayDataFile}
\alias{setAttributesBy.AromaMicroarrayDataSet}
\alias{setAttributesBySampleAnnotationFile}
\alias{setAttributesBySampleAnnotationFile.AromaMicroarrayDataSet}
\alias{setAttributesBySampleAnnotationSet}
\alias{setAttributesBySampleAnnotationSet.AromaMicroarrayDataSet}
\alias{setAttributesByTags.AromaMicroarrayDataFile}
\alias{setAttributeXY}
\alias{setAttributeXY.AromaMicroarrayDataFile}
\alias{stextLabel}
\alias{stextLabel.AromaMicroarrayDataFile}
\alias{stextLabels}
\alias{stextLabels.AromaMicroarrayDataFile}
\alias{stextSize}
\alias{stextSize.AromaMicroarrayDataFile}
\alias{allocate.AromaUgpFile}
\alias{allocateFromUnitNamesFile}
\alias{allocateFromUnitNamesFile.AromaUnitTabularBinaryFile}
\alias{AromaPlatformInterface}
\alias{AromaUgpFile}
\alias{as.character.AromaUnitTabularBinaryFile}
\alias{byChipType}
\alias{byChipType.AromaUnitTabularBinaryFile}
\alias{byName.AromaPlatform}
\alias{clearCache.AromaUnitTabularBinaryFile}
\alias{findByChipType}
\alias{findByChipType.AromaUnitTabularBinaryFile}
\alias{findUnitNamesFile}
\alias{findUnitNamesFile.AromaPlatform}
\alias{fromChipType}
\alias{fromChipType.AromaUnitTabularBinaryFile}
\alias{getAromaPlatform}
\alias{getAromaPlatform.AromaPlatformInterface}
\alias{getAromaUgpFile}
\alias{getAromaUgpFile.AromaPlatform}
\alias{getChipType.AromaUnitTabularBinaryFile}
\alias{getChipType.UnitNamesInterface}
\alias{getChromosomes.AromaUgpFile}
\alias{getColumnNames.AromaUgpFile}
\alias{getFilenameExtension}
\alias{getFilenameExtension.AromaUgpFile}
\alias{getFilenameExtension.AromaUnitTabularBinaryFile}
\alias{getGenomeVersion}
\alias{getGenomeVersion.AromaUgpFile}
\alias{getName.AromaPlatform}
\alias{getPlatform.AromaUnitTabularBinaryFile}
\alias{getPlatform.UnitNamesInterface}
\alias{getUnitNames}
\alias{getUnitNames.UnitNamesInterface}
\alias{getUnitNamesFile}
\alias{getUnitNamesFile.AromaPlatform}
\alias{getUnitNamesFile.AromaUnitTabularBinaryFile}
\alias{getUnitsAt}
\alias{getUnitsAt.AromaUgpFile}
\alias{importFromGenericTabularFile}
\alias{importFromGenericTabularFile.AromaUgpFile}
\alias{indexOfUnits}
\alias{indexOfUnits.AromaUnitTabularBinaryFile}
\alias{nbrOfUnits}
\alias{nbrOfUnits.AromaUnitTabularBinaryFile}
\alias{nbrOfUnits.UnitNamesInterface}
\alias{readDataFrame.AromaUgpFile}
\alias{clearCache.AromaUgpFile}
\alias{getPositions}
\alias{getPositions.AromaUgpFile}
\alias{getUnitsOnChromosome}
\alias{getUnitsOnChromosome.AromaUgpFile}
\alias{cnRange}
\alias{cnRange.RawCopyNumbers}
\alias{getChromosome.RawCopyNumbers}
\alias{xMax}
\alias{xMax.RawCopyNumbers}
\alias{xMin}
\alias{xMin.RawCopyNumbers}
\alias{xRange}
\alias{xRange.RawCopyNumbers}
\alias{xSeq}
\alias{xSeq.RawCopyNumbers}
\alias{as.character.AromaTransform}
\alias{clearCache.AromaTransform}
\alias{getAsteriskTags.AromaTransform}
\alias{getInputDataSet}
\alias{getOutputDataSet}
\alias{getOutputFiles}
\alias{getOutputFiles.AromaTransform}
\alias{getParameters.AromaTransform}
\alias{getParametersAsString}
\alias{getParametersAsString.AromaTransform}
\alias{getRootPath}
\alias{getRootPath.AromaTransform}
\alias{nbrOfArrays.AromaMicroarrayDataSet}
\alias{process}
\alias{setTags.AromaTransform}
\alias{getChipTypes.AromaMicroarrayDataSetTuple}
\alias{getPlatform.AromaPlatformInterface}
\alias{hasColumnHeader}
\alias{hasColumnHeader.TabularTextFile}
\alias{readRawHeader}
\alias{readRawHeader.TabularTextFile}
\alias{readTableHeader}
\alias{readTableHeader.default}
\alias{allocate.AromaCellSequenceFile}
\alias{allocate.AromaMicroarrayTabularBinaryFile}
\alias{AromaCellSequenceFile}
\alias{as.character.AromaMicroarrayTabularBinaryFile}
\alias{byChipType.AromaMicroarrayTabularBinaryFile}
\alias{countBases}
\alias{countBases.AromaCellSequenceFile}
\alias{countBasesInternal}
\alias{countBasesInternal.AromaCellSequenceFile}
\alias{findByChipType.AromaMicroarrayTabularBinaryFile}
\alias{getChipType.AromaMicroarrayTabularBinaryFile}
\alias{getColumnNames.AromaCellSequenceFile}
\alias{getFilenameExtension.AromaCellSequenceFile}
\alias{getFilenameExtension.AromaMicroarrayTabularBinaryFile}
\alias{getPlatform.AromaMicroarrayTabularBinaryFile}
\alias{getProbeLength}
\alias{getProbeLength.AromaCellSequenceFile}
\alias{importFrom}
\alias{importFrom.AromaTabularBinaryFile}
\alias{isMissing}
\alias{isMissing.AromaCellSequenceFile}
\alias{nbrOfCells}
\alias{nbrOfCells.AromaCellTabularBinaryFile}
\alias{readNeighborSequenceMatrix}
\alias{readNeighborSequenceMatrix.AromaCellSequenceFile}
\alias{readPairSequenceMatrix}
\alias{readPairSequenceMatrix.AromaCellSequenceFile}
\alias{readSequenceMatrix}
\alias{readSequenceMatrix.AromaCellSequenceFile}
\alias{readSequences}
\alias{readSequences.AromaCellSequenceFile}
\alias{remap}
\alias{remap.default}
\alias{updateSequenceMatrix}
\alias{updateSequenceMatrix.AromaCellSequenceFile}
\alias{updateSequences}
\alias{updateSequences.AromaCellSequenceFile}
\alias{readTargetStrands}
\alias{readTargetStrands.AromaCellSequenceFile}
\alias{updateTargetStrands}
\alias{updateTargetStrands.AromaCellSequenceFile}
\alias{fitProbePositionEffects}
\alias{fitProbePositionEffects.numeric}
\alias{getEffects}
\alias{getEffects.ProbePositionEffects}
\alias{getProbePositionEffectDesignMatrix}
\alias{getProbePositionEffectDesignMatrix.character}
\alias{getProbePositionEffectDesignMatrix.raw}
\alias{plot.ProbePositionEffects}
\alias{predict.ProbePositionEffects}
\alias{setFullNamesTranslator}
\alias{text.ProbePositionEffects}
\alias{update2}
\alias{update2.GenericDataFileSet}
\alias{getFullNameTranslator.GenericDataFileSet}
\alias{setFullName}
\alias{setFullName.GenericDataFile}
\alias{setFullName.GenericDataFileSet}
\alias{setFullNamesTranslator.GenericDataFileSet}
\alias{setFullNameTranslator.GenericDataFileSet}
\alias{setName}
\alias{setName.GenericDataFile}
\alias{setName.GenericDataFileSet}
\alias{translateFullName.GenericDataFileSet}
\alias{getChipType.UnitNamesFile}
\alias{getPlatform.UnitNamesFile}
\alias{getUnitNames.UnitNamesFile}
\alias{nbrOfUnits.UnitNamesFile}
\alias{dimension.TabularTextFile}
\alias{getCreatedOn}
\alias{getCreatedOn.GenericDataFile}
\alias{getLastAccessedOn}
\alias{getLastAccessedOn.GenericDataFile}
\alias{getLastModifiedOn}
\alias{getLastModifiedOn.GenericDataFile}
\alias{barSequence}
\alias{barSequence.ProbePositionEffects}
\alias{pointsSequence}
\alias{pointsSequence.ProbePositionEffects}
\alias{textSequence}
\alias{textSequence.ProbePositionEffects}
\alias{backtransformMultiDimensionalCone}
\alias{backtransformMultiDimensionalCone.matrix}
\alias{fitMultiDimensionalCone}
\alias{fitGenotypeConeByExpectile}
\alias{fitGenotypeConeByExpectile.matrix}
\alias{fitGenotypeConeBySfit}
\alias{getSnpNucleotides}
\alias{getSnpNucleotides.AromaCellSequenceFile}
\alias{getSnpPositions}
\alias{getSnpPositions.AromaCellSequenceFile}
\alias{getSnpShifts}
\alias{getSnpShifts.AromaCellSequenceFile}
\alias{groupBySnpNucleotides}
\alias{groupBySnpNucleotides.AromaCellSequenceFile}
\alias{binScatter}
\alias{binScatter.matrix}
\alias{plot.BinnedScatter}
\alias{points.BinnedScatter}
\alias{reorder.BinnedScatter}
\alias{subsample}
\alias{subsample.BinnedScatter}
\alias{subset.BinnedScatter}
\alias{getMaxLengthRepeats.AromaCellSequenceFile}
\alias{allocate.AromaSignalBinaryFile}
\alias{as.character.AromaSignalBinaryFile}
\alias{byName.AromaSignalBinarySet}
\alias{findByName.AromaSignalBinarySet}
\alias{fromFile.AromaSignalBinaryFile}
\alias{getAromaUgpFile.AromaSignalBinaryFile}
\alias{getAromaUgpFile.AromaSignalBinarySet}
\alias{getChipType.AromaSignalBinaryFile}
\alias{getChipType.AromaSignalBinarySet}
\alias{getFilenameExtension.AromaSignalBinaryFile}
\alias{getMaxLengthRepeats}
\alias{getPlatform.AromaSignalBinaryFile}
\alias{getPlatform.AromaSignalBinarySet}
\alias{getUnitNamesFile.AromaSignalBinaryFile}
\alias{nbrOfUnits.AromaSignalBinaryFile}
\alias{validate.AromaSignalBinarySet}
\alias{byName.AromaFractionBCnBinarySet}
\alias{byName.AromaTotalCnBinarySet}
\alias{extractRawCopyNumbers.AromaTotalCnBinaryFile}
\alias{allocate.AromaGenotypeCallFile}
\alias{byName.AromaGenotypeCallSet}
\alias{extractCalls}
\alias{extractCalls.AromaGenotypeCallFile}
\alias{extractCalls.AromaGenotypeCallSet}
\alias{extractConfidenceScores}
\alias{extractConfidenceScores.AromaGenotypeCallFile}
\alias{extractConfidenceScores.AromaGenotypeCallSet}
\alias{findUnitsTodo}
\alias{findUnitsTodo.AromaGenotypeCallFile}
\alias{findUnitsTodo.AromaGenotypeCallSet}
\alias{fromFiles.AromaGenotypeCallSet}
\alias{allocate.AromaUnitCallFile}
\alias{allocate.AromaUnitGenotypeCallFile}
\alias{allocate.AromaUnitSignalBinaryFile}
\alias{allocateFromUnitNamesFile.AromaUnitSignalBinaryFile}
\alias{as.character.AromaUnitSignalBinaryFile}
\alias{byName.AromaUnitCallSet}
\alias{byName.AromaUnitFracBCnBinarySet}
\alias{byName.AromaUnitSignalBinarySet}
\alias{byName.AromaUnitTotalCnBinarySet}
\alias{extractCallArray}
\alias{extractCallArray.AromaUnitCallFile}
\alias{extractCallArray.AromaUnitCallSet}
\alias{extractCalls.AromaUnitCallFile}
\alias{extractCalls.AromaUnitCallSet}
\alias{extractGenotypeMatrix}
\alias{extractGenotypeMatrix.AromaUnitCallSet}
\alias{extractGenotypeMatrix.AromaUnitGenotypeCallFile}
\alias{extractGenotypes}
\alias{extractGenotypes.AromaUnitGenotypeCallFile}
\alias{extractGenotypes.AromaUnitGenotypeCallSet}
\alias{extractMatrix.AromaUnitCallFile}
\alias{extractRawCopyNumbers.AromaUnitTotalCnBinaryFile}
\alias{findByName.AromaUnitSignalBinarySet}
\alias{findUnitsTodo.AromaUnitCallFile}
\alias{findUnitsTodo.AromaUnitCallSet}
\alias{fromFile.AromaUnitSignalBinaryFile}
\alias{fromFiles.AromaUnitCallSet}
\alias{fromFiles.AromaUnitGenotypeCallSet}
\alias{getAromaUgpFile.AromaUnitSignalBinaryFile}
\alias{getAromaUgpFile.AromaUnitSignalBinarySet}
\alias{getChipType.AromaUnitSignalBinaryFile}
\alias{getChipType.AromaUnitSignalBinarySet}
\alias{getFilenameExtension.AromaUnitSignalBinaryFile}
\alias{getPlatform.AromaUnitSignalBinaryFile}
\alias{getPlatform.AromaUnitSignalBinarySet}
\alias{getUnitNamesFile.AromaUnitSignalBinaryFile}
\alias{nbrOfUnits.AromaUnitSignalBinaryFile}
\alias{updateGenotypes}
\alias{updateGenotypes.AromaUnitGenotypeCallFile}
\alias{validate.AromaUnitSignalBinarySet}
\alias{byName.AromaUnitGenotypeCallSet}
\alias{findByName.AromaUnitCallSet}
\alias{extractMatrix.AromaUnitSignalBinaryFile}
\alias{isHeterozygote}
\alias{isHeterozygote.AromaUnitGenotypeCallFile}
\alias{isHomozygote}
\alias{isHomozygote.AromaUnitGenotypeCallFile}
\alias{getAromaUgpFile.UnitNamesFile}
\alias{shakyText.default}
\alias{shakyText}
\alias{binnedSmoothing.RawCopyNumbers}
\alias{extractSubset}
\alias{extractSubset.RawCopyNumbers}
\alias{getPositions.RawCopyNumbers}
\alias{kernelSmoothing.RawCopyNumbers}
\alias{binnedSmoothingByState}
\alias{binnedSmoothingByState.SegmentedCopyNumbers}
\alias{getStateColors}
\alias{getStateColors.SegmentedCopyNumbers}
\alias{getStates}
\alias{getStates.SegmentedCopyNumbers}
\alias{kernelSmoothingByState}
\alias{kernelSmoothingByState.SegmentedCopyNumbers}
\alias{plot.SegmentedCopyNumbers}
\alias{points.SegmentedCopyNumbers}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
