%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaTransform.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.AromaTransform}
\alias{process.AromaTransform}
\alias{AromaTransform.process}
\alias{process,AromaTransform-method}

\title{Processes the data set}

\description{
 Processes the data set.
}

\usage{
\method{process}{AromaTransform}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{double}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AromaTransform}}.
}
\keyword{internal}
\keyword{methods}
