%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SegmentedGenomicSignalsInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{SegmentedGenomicSignalsInterface}
\docType{class}
\alias{SegmentedGenomicSignalsInterface}


\title{The SegmentedGenomicSignalsInterface class interface}

\description{
 Package:  aroma.core \cr
\bold{Class SegmentedGenomicSignalsInterface}\cr

\code{\link[R.oo]{Interface}}\cr
\code{~~|}\cr
\code{~~+--}\code{SegmentedGenomicSignalsInterface}\cr

\bold{Directly known subclasses:}\cr
\link[aroma.core]{SegmentedAlleleBFractions}, \link[aroma.core]{SegmentedCopyNumbers}\cr

public class \bold{SegmentedGenomicSignalsInterface}\cr
extends \link[R.oo]{Interface}\cr


}

\usage{
SegmentedGenomicSignalsInterface(...)
}

\arguments{
  \item{...}{Not used.}
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{as.data.frame} \tab  -\cr
 \tab \code{binnedSmoothingByState} \tab  -\cr
 \tab \code{extractSubsetByState} \tab  -\cr
 \tab \code{getStateColorMap} \tab  -\cr
 \tab \code{getStateColors} \tab  -\cr
 \tab \code{getStates} \tab  -\cr
 \tab \code{getUniqueStates} \tab  -\cr
 \tab \code{kernelSmoothingByState} \tab  -\cr
 \tab \code{plot} \tab  -\cr
 \tab \code{points} \tab  -\cr
 \tab \code{setStateColorMap} \tab  -\cr
 \tab \code{setStates} \tab  -\cr
}


 \bold{Methods inherited from Interface}:\cr
extend, print, uses


}

\author{Henrik Bengtsson}
\keyword{classes}
