%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Explorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setArrays.Explorer}
\alias{setArrays.Explorer}
\alias{Explorer.setArrays}
\alias{setArrays,Explorer-method}

\title{Sets the arrays}

\description{
 Sets the arrays to be processed by the explorer.
}

\usage{
\method{setArrays}{Explorer}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Explorer}}.
}
\keyword{internal}
\keyword{methods}
