%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitAnnotationDataFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{UnitAnnotationDataFile}
\docType{class}
\alias{UnitAnnotationDataFile}


 \title{The UnitAnnotationDataFile interface class}

 \description{
  Package:  aroma.core \cr
\bold{Class UnitAnnotationDataFile}\cr

\code{\link[R.oo]{Interface}}\cr
\code{~~|}\cr
\code{~~+--}\code{UnitAnnotationDataFile}\cr

\bold{Directly known subclasses:}\cr
\link[aroma.core]{AromaUflFile}, \link[aroma.core]{AromaUgpFile}, \emph{\link[aroma.core]{AromaUnitChromosomeTabularBinaryFile}}, \emph{\link[aroma.core]{AromaUnitTabularBinaryFile}}, \link[aroma.core]{AromaUnitTypesFile}, \link[aroma.core]{TextUnitNamesFile}, \emph{\link[aroma.core]{UnitNamesFile}}, \emph{\link[aroma.core]{UnitTypesFile}}\cr

public static class \bold{UnitAnnotationDataFile}\cr
extends \link[R.oo]{Interface}\cr



  A UnitAnnotationDataFile provides methods for querying certain types
  of chip type annotation data by units.
 }

 \usage{UnitAnnotationDataFile(...)}

 \arguments{
   \item{...}{Arguments passed to \code{\link[R.oo]{Interface}}.}
 }

 \section{Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{byChipType} \tab  -\cr
 \tab \code{getAromaUflFile} \tab  -\cr
 \tab \code{getAromaUgpFile} \tab  -\cr
 \tab \code{getChipType} \tab  -\cr
 \tab \code{getPlatform} \tab  -\cr
 \tab \code{nbrOfUnits} \tab  -\cr
}


 \bold{Methods inherited from Interface}:\cr
extend, print, uses


 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
