%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomeExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{updateSetupExplorerFile.ChromosomeExplorer}
\alias{updateSetupExplorerFile.ChromosomeExplorer}
\alias{ChromosomeExplorer.updateSetupExplorerFile}
\alias{updateSetupExplorerFile,ChromosomeExplorer-method}

\title{Updates the Javascript file}

\description{
 Updates the Javascript file.
}

\usage{
\method{updateSetupExplorerFile}{ChromosomeExplorer}(this, ..., verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns (invisibly) the pathname to the samples.js file.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomeExplorer}}.
}
\keyword{internal}
\keyword{methods}
