%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaTabularBinaryFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeFooter.AromaTabularBinaryFile}
\alias{writeFooter.AromaTabularBinaryFile}
\alias{AromaTabularBinaryFile.writeFooter}
\alias{writeFooter,AromaTabularBinaryFile-method}

\title{Writes a named nested list to the file footer in XML format}

\description{
 Writes a named nested list to the file footer in XML format.
}

\usage{
\method{writeFooter}{AromaTabularBinaryFile}(this, footer, ...)
}

\arguments{
  \item{footer}{A named \code{\link[base]{list}} structure.}
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[aroma.core:readFooter.AromaTabularBinaryFile]{*readFooter}()}.
  For more information see \code{\link{AromaTabularBinaryFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
