%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{aroma.core-package}
\alias{aroma.core-package}
\alias{aroma.core}
\docType{package}

\title{Package aroma.core}


 \description{
   This package contains core methods and classes used by         higher-level aroma.* packages part of the Aroma Project, e.g.         aroma.affymetrix and aroma.cn.  Its API is mostly of interest 	to developers and is rarely used by end users directly.

   This package is a support package for \pkg{aroma.affymetrix}.
   This package should be considered to be in an alpha or beta phase.
   You should expect the API to be changing over time.
   For practical purposes, consider this package a private package.
 }

 \section{Installation and updates}{

   This package is available on CRAN, i.e. to install do
   \code{install.packages("aroma.core")}.
 }

 \author{Henrik Bengtsson}

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.

   The development code of the packages is under a private licence
   (where applicable) and patches sent to the author fall under the
   latter license, but will be, if incorporated, released under the
   "release" license above.
 }

 \references{
 [1] H. Bengtsson, \emph{The R.oo package - Object-Oriented Programming with References Using Standard R Code}, In Kurt Hornik, Friedrich Leisch and Achim Zeileis, editors, Proceedings of the 3rd International Workshop on Distributed Statistical Computing (DSC 2003), March 20-22, Vienna, Austria. \url{http://www.ci.tuwien.ac.at/Conferences/DSC-2003/Proceedings/}
 \cr
 }
\keyword{package}
