%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParametersInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getParametersAsString.ParametersInterface}
\alias{getParametersAsString.ParametersInterface}
\alias{ParametersInterface.getParametersAsString}
\alias{getParametersAsString.ParametersInterface}
\alias{getParametersAsString,ParametersInterface-method}


\title{Gets the parameters as character}

\description{
 Gets the parameters as character.
}

\usage{\method{getParametersAsString}{ParametersInterface}(this, ..., collapse=NULL)}

\arguments{
 \item{...}{Arguments passed to \code{\link[aroma.core:getParameters.ParametersInterface]{*getParameters}()}.}
 \item{collapse}{(optional) A \code{\link[base]{character}} string used to collapse the
   individual parameter strings.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ParametersInterface}}.
}
\keyword{internal}
\keyword{methods}
