%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaTransform.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.AromaTransform}
\alias{getPath.AromaTransform}
\alias{AromaTransform.getPath}
\alias{getPath,AromaTransform-method}

\title{Gets the path of the output directory}

\description{
 Gets the path of the output directory.
}

\usage{
\method{getPath}{AromaTransform}(this, create=TRUE, ...)
}

\arguments{
  \item{create}{If \code{\link[base:logical]{TRUE}}, the path is created, otherwise not.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\details{
  Windows Shortcut links are recognized.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AromaTransform}}.
}
\keyword{internal}
\keyword{methods}
