%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaTransform.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isDone.AromaTransform}
\alias{isDone.AromaTransform}
\alias{AromaTransform.isDone}
\alias{isDone,AromaTransform-method}

\title{Checks if the data set is processed or not}

\description{
 Checks if the data set is processed or not.
}

\usage{
\method{isDone}{AromaTransform}(this, ...)
}

\arguments{
  \item{...}{Arguments passed to \code{\link[aroma.core:findFilesTodo.AromaTransform]{*findFilesTodo}()}.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the data set is processed, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AromaTransform}}.
}
\keyword{internal}
\keyword{methods}
