%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaMicroarrayDataSetTuple.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfChipTypes.AromaMicroarrayDataSetTuple}
\alias{nbrOfChipTypes.AromaMicroarrayDataSetTuple}
\alias{AromaMicroarrayDataSetTuple.nbrOfChipTypes}
\alias{nbrOfChipTypes,AromaMicroarrayDataSetTuple-method}

\title{Gets the number of chip types}

\description{
 Gets the number of chip types used in the model.
}

\usage{
\method{nbrOfChipTypes}{AromaMicroarrayDataSetTuple}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[aroma.core:getChipTypes.AromaMicroarrayDataSetTuple]{*getChipTypes}()}.
  For more information see \code{\link{AromaMicroarrayDataSetTuple}}.
}
\keyword{internal}
\keyword{methods}
