%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{[.AromaTabularBinaryFile}
\alias{[[.AromaTabularBinaryFile}
\alias{[<-.AromaTabularBinaryFile}
\alias{addBy}
\alias{addBy.RawGenomicSignals}
\alias{addLocusFields}
\alias{addLocusFields.RawGenomicSignals}
\alias{allocate}
\alias{allocate.AromaCellSequenceFile}
\alias{allocate.AromaMicroarrayTabularBinaryFile}
\alias{allocate.AromaUgpFile}
\alias{allocate.AromaUnitCallFile}
\alias{allocate.AromaUnitChromosomeTabularBinaryFile}
\alias{allocate.AromaUnitGenotypeCallFile}
\alias{allocate.AromaUnitSignalBinaryFile}
\alias{allocateFromUnitNamesFile}
\alias{allocateFromUnitNamesFile.AromaUnitSignalBinaryFile}
\alias{allocateFromUnitNamesFile.AromaUnitTabularBinaryFile}
\alias{append.RawGenomicSignals}
\alias{apply}
\alias{apply.default}
\alias{apply.SampleAnnotationFile}
\alias{applyBinaryOperator}
\alias{applyBinaryOperator.RawGenomicSignals}
\alias{applyRegions}
\alias{applyRegions.CopyNumberOutliers}
\alias{applyRows}
\alias{applyRows.CopyNumberRegions}
\alias{AromaCellSequenceFile}
\alias{AromaCore}
\alias{AromaPackage}
\alias{AromaPlatformInterface}
\alias{aromaSettings}
\alias{AromaSettings}
\alias{AromaUgpFile}
\alias{AromaUnitChromosomeTabularBinaryFile}
\alias{AromaUnitTotalCnBinaryFileList}
\alias{as.character.AromaMicroarrayDataSetTuple}
\alias{as.character.AromaMicroarrayTabularBinaryFile}
\alias{as.character.AromaTabularBinaryFile}
\alias{as.character.AromaTransform}
\alias{as.character.AromaUnitSignalBinaryFile}
\alias{as.character.CopyNumberOutliers}
\alias{as.character.CopyNumberRegions}
\alias{as.character.RawGenomicSignals}
\alias{as.character.TextUnitNamesFile}
\alias{as.data.frame.CopyNumberOutliers}
\alias{as.data.frame.CopyNumberRegions}
\alias{as.data.frame.RawCopyNumbers}
\alias{as.data.frame.RawGenomicSignals}
\alias{as.data.frame.SegmentedGenomicSignalsInterface}
\alias{as.GrayscaleImage}
\alias{as.TrueColorImage}
\alias{as.TrueColorImage.Image}
\alias{as.TrueColorImage.matrix}
\alias{asMatrixOfFiles}
\alias{asMatrixOfFiles.AromaMicroarrayDataSetTuple}
\alias{asThis}
\alias{asThis.BasicObject}
\alias{asThis.Object}
\alias{backtransformGenotypeCone}
\alias{backtransformGenotypeCone.matrix}
\alias{backtransformMultiDimensionalCone}
\alias{backtransformMultiDimensionalCone.matrix}
\alias{barSequence}
\alias{barSequence.ProbePositionEffects}
\alias{binnedSmoothing.RawGenomicSignals}
\alias{binnedSmoothingByState}
\alias{binnedSmoothingByState.SegmentedGenomicSignalsInterface}
\alias{binScatter}
\alias{binScatter.matrix}
\alias{byChipType}
\alias{byChipType.AromaMicroarrayTabularBinaryFile}
\alias{byChipType.AromaUnitTabularBinaryFile}
\alias{byChipType.UnitNamesFile}
\alias{byName.AromaPlatform}
\alias{byName.AromaUnitFracBCnBinarySet}
\alias{byName.AromaUnitGenotypeCallSet}
\alias{byName.AromaUnitSignalBinarySet}
\alias{byName.AromaUnitTotalCnBinarySet}
\alias{byPath}
\alias{byPath.AromaMicroarrayDataSetTuple}
\alias{calcMargins}
\alias{calcMargins.matrix}
\alias{clearCache.AromaTabularBinaryFile}
\alias{clearCache.AromaTransform}
\alias{clearCache.AromaUnitChromosomeTabularBinaryFile}
\alias{clearCache.AromaUnitTabularBinaryFile}
\alias{clearCache.TextUnitNamesFile}
\alias{cnRange}
\alias{cnRange.RawCopyNumbers}
\alias{colGaussianSmoothing}
\alias{colGaussianSmoothing.matrix}
\alias{colMeans}
\alias{colMeans.AromaTabularBinaryFile}
\alias{colMeans.default}
\alias{colMedians.AromaTabularBinaryFile}
\alias{colorize}
\alias{colorize.Image}
\alias{colStats}
\alias{colStats.AromaTabularBinaryFile}
\alias{colSums}
\alias{colSums.AromaTabularBinaryFile}
\alias{colSums.default}
\alias{convertTable}
\alias{convertTable.default}
\alias{CopyNumberOutliers}
\alias{CopyNumberRegions}
\alias{countBases}
\alias{countBases.AromaCellSequenceFile}
\alias{countBasesInternal}
\alias{countBasesInternal.AromaCellSequenceFile}
\alias{dimnames<-.AromaTabularBinaryFile}
\alias{display}
\alias{display.Image}
\alias{divideBy}
\alias{divideBy.RawGenomicSignals}
\alias{drawCnRegions}
\alias{drawCnRegions.DNAcopy}
\alias{drawCnRegions.HaarSeg}
\alias{drawCnRegions.profileCGH}
\alias{drawCytoband}
\alias{drawCytoband.profileCGH}
\alias{drawCytoband2}
\alias{drawCytoband2.default}
\alias{drawLevels}
\alias{drawLevels.CopyNumberOutliers}
\alias{drawLevels.CopyNumberRegions}
\alias{estimateStandardDeviation}
\alias{exportTotalCnRatioSet}
\alias{exportTotalCnRatioSet.AromaUnitTotalCnBinarySet}
\alias{extractByChromosome}
\alias{extractByChromosome.AromaUnitChromosomeTabularBinaryFile}
\alias{extractCallArray}
\alias{extractCallArray.AromaUnitCallFile}
\alias{extractCallArray.AromaUnitCallSet}
\alias{extractCalls}
\alias{extractCalls.AromaUnitCallFile}
\alias{extractCalls.AromaUnitCallSet}
\alias{extractCNRs}
\alias{extractCNRs.default}
\alias{extractCopyNumberRegions}
\alias{extractCopyNumberRegions.default}
\alias{extractCopyNumberRegions.DNAcopy}
\alias{extractCopyNumberRegions.HaarSeg}
\alias{extractCopyNumberRegions.profileCGH}
\alias{extractDataForSegmentation}
\alias{extractDataForSegmentation.RawGenomicSignals}
\alias{extractGenotypeMatrix}
\alias{extractGenotypeMatrix.AromaUnitCallSet}
\alias{extractGenotypeMatrix.AromaUnitGenotypeCallFile}
\alias{extractGenotypes}
\alias{extractGenotypes.AromaUnitGenotypeCallFile}
\alias{extractGenotypes.AromaUnitGenotypeCallSet}
\alias{extractMatrix.AromaUnitCallFile}
\alias{extractMatrix.AromaUnitSignalBinaryFile}
\alias{extractMergedRawCopyNumbers}
\alias{extractMergedRawCopyNumbers.AromaUnitTotalCnBinaryFileList}
\alias{extractRawAlleleBFractions}
\alias{extractRawAlleleBFractions.default}
\alias{extractRawCNs}
\alias{extractRawCNs.default}
\alias{extractRawCopyNumbers}
\alias{extractRawCopyNumbers.AromaUnitTotalCnBinaryFile}
\alias{extractRawCopyNumbers.AromaUnitTotalCnBinaryFileList}
\alias{extractRawCopyNumbers.default}
\alias{extractRawCopyNumbers.DNAcopy}
\alias{extractRawCopyNumbers.HaarSeg}
\alias{extractRawCopyNumbers.profileCGH}
\alias{extractRawGenomicSignals}
\alias{extractRawGenomicSignals.AromaUnitTotalCnBinaryFileList}
\alias{extractRawGenomicSignals.default}
\alias{extractRegion}
\alias{extractRegion.RawGenomicSignals}
\alias{extractSubset}
\alias{extractSubset.RawGenomicSignals}
\alias{extractSubsetByState}
\alias{extractSubsetByState.SegmentedGenomicSignalsInterface}
\alias{findByChipType}
\alias{findByChipType.AromaMicroarrayTabularBinaryFile}
\alias{findByChipType.TextUnitNamesFile}
\alias{findByName.AromaUnitCallSet}
\alias{findByName.AromaUnitSignalBinarySet}
\alias{findPngDevice}
\alias{findPngDevice.default}
\alias{findSAFs}
\alias{findSAFs.SampleAnnotationSet}
\alias{findUnitNamesFile}
\alias{findUnitNamesFile.AromaPlatform}
\alias{findUnitsTodo}
\alias{findUnitsTodo.AromaUnitCallFile}
\alias{findUnitsTodo.AromaUnitCallSet}
\alias{fit2d}
\alias{fit2d.matrix}
\alias{fitGenotypeCone}
\alias{fitGenotypeConeByExpectile}
\alias{fitGenotypeConeByExpectile.matrix}
\alias{fitGenotypeConeBySfit}
\alias{fitMultiDimensionalCone}
\alias{fitProbePositionEffects}
\alias{fitProbePositionEffects.numeric}
\alias{fitSplineBlockPolish}
\alias{fitSplineBlockPolish.matrix}
\alias{fitWHLAPLM}
\alias{fitWHLAPLM.matrix}
\alias{fitWHRCModel}
\alias{fitWHRCModel.matrix}
\alias{fitWLAPLM}
\alias{fitWLAPLM.matrix}
\alias{fitWRCModel}
\alias{fitWRCModel.matrix}
\alias{fitWRMA}
\alias{fitWRMA.matrix}
\alias{fixSearchPath}
\alias{fixSearchPath.AromaCore}
\alias{fixSearchPath.AromaPackage}
\alias{fixSearchPathInternal}
\alias{fixSearchPathInternal.AromaPackage}
\alias{fromFile.AromaUnitSignalBinaryFile}
\alias{fromFiles.AromaUnitCallSet}
\alias{fromFiles.AromaUnitGenotypeCallSet}
\alias{fromPath}
\alias{fromPath.SampleAnnotationFile}
\alias{fromPath.SampleAnnotationSet}
\alias{gaussianSmoothing}
\alias{gaussianSmoothing.matrix}
\alias{gaussianSmoothing.numeric}
\alias{gaussianSmoothing.RawGenomicSignals}
\alias{gaussKernel}
\alias{GenericSummary}
\alias{getAlias.AromaMicroarrayDataSetTuple}
\alias{getAromaPlatform}
\alias{getAromaPlatform.AromaPlatformInterface}
\alias{getAromaUgpFile}
\alias{getAromaUgpFile.AromaPlatform}
\alias{getAromaUgpFile.AromaUnitSignalBinaryFile}
\alias{getAromaUgpFile.AromaUnitSignalBinarySet}
\alias{getAromaUgpFile.UnitNamesFile}
\alias{getArrays}
\alias{getArrayTuple}
\alias{getAsteriskTags}
\alias{getAsteriskTags.AromaMicroarrayDataSetTuple}
\alias{getAsteriskTags.AromaTransform}
\alias{getAttributeXY}
\alias{getAttributeXY.AromaMicroarrayDataFile}
\alias{getBytesPerColumn}
\alias{getBytesPerColumn.AromaTabularBinaryFile}
\alias{getChipType}
\alias{getChipType.AromaMicroarrayDataFile}
\alias{getChipType.AromaMicroarrayDataSet}
\alias{getChipType.AromaMicroarrayTabularBinaryFile}
\alias{getChipType.AromaUnitSignalBinaryFile}
\alias{getChipType.AromaUnitSignalBinarySet}
\alias{getChipType.TextUnitNamesFile}
\alias{getChipType.UnitNamesFile}
\alias{getChipTypes}
\alias{getChipTypes.AromaMicroarrayDataSetTuple}
\alias{getChromosome}
\alias{getChromosome.Arguments}
\alias{getChromosome.RawGenomicSignals}
\alias{getChromosomes}
\alias{getChromosomes.Arguments}
\alias{getChromosomes.AromaUnitChromosomeTabularBinaryFile}
\alias{getCn}
\alias{getCn.RawCopyNumbers}
\alias{getCNs}
\alias{getCNs.RawCopyNumbers}
\alias{getColClasses}
\alias{getColClasses.AromaTabularBinaryFile}
\alias{getColumnNames.AromaCellSequenceFile}
\alias{getColumnNames.AromaTabularBinaryFile}
\alias{getColumnNames.AromaUgpFile}
\alias{getColumnNames.AromaUnitChromosomeTabularBinaryFile}
\alias{getCommonListElements}
\alias{getCommonPrefix}
\alias{getDeviceResolution}
\alias{getDeviceResolution.default}
\alias{getEffects}
\alias{getEffects.ProbePositionEffects}
\alias{getFilenameExtension}
\alias{getFilenameExtension.AromaCellSequenceFile}
\alias{getFilenameExtension.AromaMicroarrayTabularBinaryFile}
\alias{getFilenameExtension.AromaUgpFile}
\alias{getFilenameExtension.AromaUnitChromosomeTabularBinaryFile}
\alias{getFilenameExtension.AromaUnitSignalBinaryFile}
\alias{getFilenameExtension.TextUnitNamesFile}
\alias{getFullName.AromaMicroarrayDataSetTuple}
\alias{getFullNames.AromaMicroarrayDataSetTuple}
\alias{getGenomeVersion}
\alias{getGenomeVersion.AromaUnitChromosomeTabularBinaryFile}
\alias{getHeaderParameters}
\alias{getHeaderParameters.TextUnitNamesFile}
\alias{getImage}
\alias{getImage.matrix}
\alias{getInputDataSet}
\alias{getLabel.AromaMicroarrayDataFile}
\alias{getListOfSets}
\alias{getListOfSets.AromaMicroarrayDataSetTuple}
\alias{getLociFields}
\alias{getLociFields.RawGenomicSignals}
\alias{getLocusFields}
\alias{getLocusFields.RawGenomicSignals}
\alias{getMaxLengthRepeats}
\alias{getMaxLengthRepeats.AromaCellSequenceFile}
\alias{getName.AromaMicroarrayDataSetTuple}
\alias{getName.AromaPlatform}
\alias{getName.RawGenomicSignals}
\alias{getNames.AromaMicroarrayDataSetTuple}
\alias{getOutputDataSet}
\alias{getOutputFiles}
\alias{getOutputFiles.AromaTransform}
\alias{getParameters.AromaTransform}
\alias{getParametersAsString}
\alias{getParametersAsString.AromaTransform}
\alias{getPatterns}
\alias{getPatterns.SampleAnnotationFile}
\alias{getPhysicalPositions}
\alias{getPhysicalPositions.RawCopyNumbers}
\alias{getPlatform}
\alias{getPlatform.AromaMicroarrayDataFile}
\alias{getPlatform.AromaMicroarrayDataSet}
\alias{getPlatform.AromaMicroarrayTabularBinaryFile}
\alias{getPlatform.AromaPlatformInterface}
\alias{getPlatform.AromaUnitSignalBinaryFile}
\alias{getPlatform.AromaUnitSignalBinarySet}
\alias{getPlatform.TextUnitNamesFile}
\alias{getPlatform.UnitNamesFile}
\alias{getPloidy}
\alias{getPloidy.AromaMicroarrayDataFile}
\alias{getPositions}
\alias{getPositions.AromaUgpFile}
\alias{getPositions.RawGenomicSignals}
\alias{getProbeLength}
\alias{getProbeLength.AromaCellSequenceFile}
\alias{getProbePositionEffectDesignMatrix}
\alias{getProbePositionEffectDesignMatrix.character}
\alias{getProbePositionEffectDesignMatrix.raw}
\alias{getRam}
\alias{getRam.AromaSettings}
\alias{getRootPath}
\alias{getRootPath.AromaTransform}
\alias{getSigma}
\alias{getSigma.RawGenomicSignals}
\alias{getSignals}
\alias{getSignals.RawGenomicSignals}
\alias{getSnpNucleotides}
\alias{getSnpNucleotides.AromaCellSequenceFile}
\alias{getSnpPositions}
\alias{getSnpPositions.AromaCellSequenceFile}
\alias{getSnpShifts}
\alias{getSnpShifts.AromaCellSequenceFile}
\alias{getStateColors}
\alias{getStateColors.SegmentedGenomicSignalsInterface}
\alias{getStates}
\alias{setStates.SegmentedGenomicSignalsInterface}
\alias{getStates.SegmentedGenomicSignalsInterface}
\alias{getTableOfArrays}
\alias{getTags.AromaMicroarrayDataSetTuple}
\alias{getUniqueStates}
\alias{getUniqueStates.SegmentedGenomicSignalsInterface}
\alias{getUnitNames}
\alias{getUnitNames.TextUnitNamesFile}
\alias{getUnitNames.UnitNamesFile}
\alias{getUnitNamesFile}
\alias{getUnitNamesFile.AromaPlatform}
\alias{getUnitNamesFile.AromaPlatformInterface}
\alias{getUnitsAt}
\alias{getUnitsAt.AromaUgpFile}
\alias{getUnitsOnChromosome}
\alias{getUnitsOnChromosome.AromaUgpFile}
\alias{getUnitsOnChromosome.AromaUnitChromosomeTabularBinaryFile}
\alias{getUnitsOnChromosomes}
\alias{getUnitsOnChromosomes.AromaUnitChromosomeTabularBinaryFile}
\alias{getVerbose.AromaSettings}
\alias{getWeights}
\alias{getWeights.RawGenomicSignals}
\alias{getXScale}
\alias{getXScale.RawGenomicSignals}
\alias{getXY}
\alias{getXY.RawGenomicSignals}
\alias{getYScale}
\alias{getYScale.RawGenomicSignals}
\alias{groupBySnpNucleotides}
\alias{groupBySnpNucleotides.AromaCellSequenceFile}
\alias{hasAttributeXY}
\alias{hasAttributeXY.AromaMicroarrayDataFile}
\alias{hasWeights}
\alias{hasWeights.RawGenomicSignals}
\alias{importFrom}
\alias{importFrom.AromaTabularBinaryFile}
\alias{importFromGenericTabularFile}
\alias{importFromGenericTabularFile.AromaUgpFile}
\alias{importFromTable}
\alias{importFromTable.FileMatrix}
\alias{indexOfArrays}
\alias{indexOfColumn}
\alias{indexOfColumn.AromaUnitChromosomeTabularBinaryFile}
\alias{indexOfUnits}
\alias{indexOfUnits.AromaUnitTabularBinaryFile}
\alias{interleave}
\alias{interleave.Image}
\alias{isHeterozygote}
\alias{isHeterozygote.AromaUnitGenotypeCallFile}
\alias{isHomozygote}
\alias{isHomozygote.AromaUnitGenotypeCallFile}
\alias{isMissing}
\alias{isMissing.AromaCellSequenceFile}
\alias{kernelSmoothing.RawGenomicSignals}
\alias{kernelSmoothingByState}
\alias{kernelSmoothingByState.SegmentedGenomicSignalsInterface}
\alias{lapply.AromaTabularBinaryFile}
\alias{library}
\alias{lines.CopyNumberOutliers}
\alias{lines.CopyNumberRegions}
\alias{lines.RawGenomicSignals}
\alias{listToXml}
\alias{listToXml.list}
\alias{locallyUnique}
\alias{locallyUnique.default}
\alias{log2abs}
\alias{log2center}
\alias{log2neg}
\alias{log2pos}
\alias{matchPatterns}
\alias{matchPatterns.SampleAnnotationFile}
\alias{matrixBlockPolish}
\alias{mergeBoxplotStats}
\alias{mergeByCommonTails}
\alias{multiplyBy}
\alias{multiplyBy.RawGenomicSignals}
\alias{nbrOfArrays}
\alias{nbrOfArrays.AromaMicroarrayDataSet}
\alias{nbrOfCells}
\alias{nbrOfCells.AromaCellTabularBinaryFile}
\alias{nbrOfChipTypes}
\alias{nbrOfColumns.AromaTabularBinaryFile}
\alias{nbrOfLoci}
\alias{nbrOfLoci.RawGenomicSignals}
\alias{nbrOfRegions}
\alias{nbrOfRegions.CopyNumberOutliers}
\alias{nbrOfRegions.CopyNumberRegions}
\alias{nbrOfRows.AromaTabularBinaryFile}
\alias{nbrOfUnits}
\alias{nbrOfUnits.AromaUnitSignalBinaryFile}
\alias{nbrOfUnits.AromaUnitTabularBinaryFile}
\alias{nbrOfUnits.TextUnitNamesFile}
\alias{nbrOfUnits.UnitNamesFile}
\alias{norm2d}
\alias{norm2d.matrix}
\alias{normalizeGenotypeCone}
\alias{normalizeGenotypeCone.matrix}
\alias{patch}
\alias{patch.AromaPackage}
\alias{plot.BinnedScatter}
\alias{plot.ProbePositionEffects}
\alias{plot.RawAlleleBFractions}
\alias{plot.RawCopyNumbers}
\alias{plot.RawGenomicSignals}
\alias{plot.SegmentedGenomicSignalsInterface}
\alias{points.BinnedScatter}
\alias{points.RawGenomicSignals}
\alias{points.SegmentedGenomicSignalsInterface}
\alias{pointsSequence}
\alias{pointsSequence.ProbePositionEffects}
\alias{predict.ProbePositionEffects}
\alias{print.GenericSummary}
\alias{process}
\alias{readColumns.AromaTabularBinaryFile}
\alias{readData}
\alias{readData.SampleAnnotationFile}
\alias{readDataFrame.AromaTabularBinaryFile}
\alias{readDataFrame.AromaUnitChromosomeTabularBinaryFile}
\alias{readDataFrame.AromaUnitSignalBinaryFile}
\alias{readDataFrame.SampleAnnotationFile}
\alias{readFooter}
\alias{readHeader}
\alias{readHeader.AromaTabularBinaryFile}
\alias{readNeighborSequenceMatrix}
\alias{readNeighborSequenceMatrix.AromaCellSequenceFile}
\alias{readPairSequenceMatrix}
\alias{readPairSequenceMatrix.AromaCellSequenceFile}
\alias{readRawFooter}
\alias{readRawFooter.AromaTabularBinaryFile}
\alias{readSequenceMatrix}
\alias{readSequenceMatrix.AromaCellSequenceFile}
\alias{readSequences}
\alias{readSequences.AromaCellSequenceFile}
\alias{readTargetStrands}
\alias{readTargetStrands.AromaCellSequenceFile}
\alias{remap}
\alias{remap.default}
\alias{reorder.BinnedScatter}
\alias{require}
\alias{rescale}
\alias{rescale.Image}
\alias{rgbTransform}
\alias{rgbTransform.Image}
\alias{SampleAnnotationFile}
\alias{SampleAnnotationSet}
\alias{segmentByCBS}
\alias{segmentByGLAD}
\alias{segmentByHaarSeg}
\alias{setAlias.AromaMicroarrayDataSetTuple}
\alias{setAttributesBy.AromaMicroarrayDataSet}
\alias{setAttributesBySampleAnnotationFile}
\alias{setAttributesBySampleAnnotationFile.AromaMicroarrayDataSet}
\alias{setAttributesBySampleAnnotationSet}
\alias{setAttributesBySampleAnnotationSet.AromaMicroarrayDataSet}
\alias{setAttributesByTags.AromaMicroarrayDataFile}
\alias{setAttributesByTags.AromaTabularBinaryFile}
\alias{setAttributeXY}
\alias{setAttributeXY.AromaMicroarrayDataFile}
\alias{setLabel.AromaMicroarrayDataFile}
\alias{setLocusFields}
\alias{setLocusFields.RawGenomicSignals}
\alias{setRam}
\alias{setRam.AromaSettings}
\alias{setSigma}
\alias{setSigma.RawGenomicSignals}
\alias{setTags.AromaMicroarrayDataSetTuple}
\alias{setTags.AromaTransform}
\alias{setVerbose}
\alias{setVerbose.AromaSettings}
\alias{setWeights}
\alias{setWeights.RawGenomicSignals}
\alias{setXScale}
\alias{setXScale.RawGenomicSignals}
\alias{setYScale}
\alias{setYScale.RawGenomicSignals}
\alias{shakyText}
\alias{shakyText.default}
\alias{signalRange}
\alias{signalRange.RawGenomicSignals}
\alias{smoothWRMA}
\alias{smoothWRMA.matrix}
\alias{smoothWSA}
\alias{smoothWSA.matrix}
\alias{sort.RawGenomicSignals}
\alias{splitByCommonTails}
\alias{sqrtabs}
\alias{sqrtcenter}
\alias{sqrtneg}
\alias{sqrtpos}
\alias{stext}
\alias{stext.default}
\alias{stextChipType}
\alias{stextChipType.character}
\alias{stextLabel}
\alias{stextLabel.AromaMicroarrayDataFile}
\alias{stextLabels}
\alias{stextLabels.AromaMicroarrayDataFile}
\alias{stextSize}
\alias{stextSize.AromaMicroarrayDataFile}
\alias{stringTree}
\alias{stringTree.character}
\alias{subsample}
\alias{subsample.BinnedScatter}
\alias{subset.AromaTabularBinaryFile}
\alias{subset.BinnedScatter}
\alias{subtractBy}
\alias{subtractBy.RawGenomicSignals}
\alias{summary.AromaTabularBinaryFile}
\alias{summary.RawGenomicSignals}
\alias{text.ProbePositionEffects}
\alias{textSequence}
\alias{textSequence.ProbePositionEffects}
\alias{TextUnitNamesFile}
\alias{updateData}
\alias{updateData.AromaTabularBinaryFile}
\alias{updateDataColumn}
\alias{updateDataColumn.AromaTabularBinaryFile}
\alias{updateGenotypes}
\alias{updateGenotypes.AromaUnitGenotypeCallFile}
\alias{updateSequenceMatrix}
\alias{updateSequenceMatrix.AromaCellSequenceFile}
\alias{updateSequences}
\alias{updateSequences.AromaCellSequenceFile}
\alias{updateTargetStrands}
\alias{updateTargetStrands.AromaCellSequenceFile}
\alias{uses}
\alias{uses.character}
\alias{validate.AromaMicroarrayDataSet}
\alias{validate.AromaUnitSignalBinarySet}
\alias{weightedMad}
\alias{weightedMean}
\alias{whatDataType}
\alias{whatDataType.default}
\alias{writeFooter}
\alias{writeRawFooter}
\alias{writeRawFooter.AromaTabularBinaryFile}
\alias{xMax}
\alias{xMax.RawGenomicSignals}
\alias{xMin}
\alias{xMin.RawGenomicSignals}
\alias{xmlToList}
\alias{xmlToList.character}
\alias{xRange}
\alias{xRange.RawGenomicSignals}
\alias{xSeq}
\alias{xSeq.RawGenomicSignals}
\alias{yellow.colors}
\alias{All}
\alias{user-level}
\alias{objects}
\alias{in}
\alias{a}
\alias{package}
\alias{should}
\alias{have}
\alias{documentation}
\alias{entries.}
\alias{calculateAverageColumnAcrossFiles}
\alias{calculateAverageColumnAcrossFiles.GenericTabularFileSet}
\alias{lapplyInChunks}
\alias{lapplyInChunks.numeric}
\alias{splitInChunks}
\alias{splitInChunks.numeric}
\alias{allocate.AromaCellPositionFile}
\alias{AromaCellPositionFile}
\alias{byChipType.AromaCellPositionFile}
\alias{getColumnNames.AromaCellPositionFile}
\alias{getFilenameExtension.AromaCellPositionFile}
\alias{isMissing.AromaCellPositionFile}
\alias{readPositions}
\alias{readPositions.AromaCellPositionFile}
\alias{updatePositions}
\alias{updatePositions.AromaCellPositionFile}
\alias{addIncludes}
\alias{addIncludes.Explorer}
\alias{addIndexFile}
\alias{addIndexFile.Explorer}
\alias{as.character.Explorer}
\alias{getAsteriskTags.Explorer}
\alias{getFullName.Explorer}
\alias{getIncludePath}
\alias{getIncludePath.Explorer}
\alias{getMainPath}
\alias{getMainPath.Explorer}
\alias{getNameOfInput}
\alias{getNameOfInput.Explorer}
\alias{getParallelSafe}
\alias{getParallelSafe.Explorer}
\alias{getPath.Explorer}
\alias{getReportPathPattern}
\alias{getReportPathPattern.Explorer}
\alias{getRootPath.Explorer}
\alias{getSampleLayerPrefix}
\alias{getSampleLayerPrefix.Explorer}
\alias{getSubname}
\alias{getSubname.Explorer}
\alias{getTagsOfInput}
\alias{getTagsOfInput.Explorer}
\alias{getTemplatePath}
\alias{getTemplatePath.Explorer}
\alias{setArrays}
\alias{setParallelSafe}
\alias{setParallelSafe.Explorer}
\alias{setReportPathPattern}
\alias{setReportPathPattern.Explorer}
\alias{setSubname}
\alias{setSubname.Explorer}
\alias{setup}
\alias{setup.Explorer}
\alias{splitByReportPathPattern}
\alias{splitByReportPathPattern.Explorer}
\alias{getArraysOfInput}
\alias{getArraysOfInput.Explorer}
\alias{extractRawAlleleBFractions.AromaUnitFracBCnBinaryFile}
\alias{extractRawGenomicSignals.AromaUnitSignalBinaryFile}
\alias{extractRawMirroredAlleleBFractions}
\alias{extractRawMirroredAlleleBFractions.default}
\alias{extractRawMirroredAlleleBFractions.RawAlleleBFractions}
\alias{plot.RawMirroredAlleleBFractions}
\alias{exportFracBDiffSet}
\alias{exportFracBDiffSet.AromaUnitFracBCnBinarySet}
\alias{extractIGV}
\alias{extractIGV.CopyNumberRegions}
\alias{getExpectedOutputFiles}
\alias{getExpectedOutputFiles.AromaTransform}
\alias{getOutputDataSet0}
\alias{getOutputDataSet0.AromaTransform}
\alias{getExpectedOutputFullnames}
\alias{getExpectedOutputFullnames.AromaTransform}
\alias{isHeterozygous}
\alias{isHeterozygous.AromaUnitGenotypeCallFile}
\alias{isHomozygous}
\alias{isHomozygous.AromaUnitGenotypeCallFile}
\alias{setStates}
\alias{getStateColorMap}
\alias{getStateColorMap.SegmentedGenomicSignalsInterface}
\alias{setStateColorMap}
\alias{setStateColorMap.SegmentedGenomicSignalsInterface}
\alias{binnedSums}
\alias{binnedSums.RawSequenceReads}
\alias{nbrOfReads}
\alias{nbrOfReads.RawSequenceReads}
\alias{plot.RawSequenceReads}
\alias{allocateFromUnitAnnotationDataFile}
\alias{allocateFromUnitAnnotationDataFile.AromaUnitSignalBinaryFile}
\alias{allocateFromUnitAnnotationDataFile.AromaUnitTabularBinaryFile}
\alias{byChipType.UnitAnnotationDataFile}
\alias{findUnitTypesFile}
\alias{findUnitTypesFile.AromaPlatform}
\alias{getAromaUgpFile.UnitAnnotationDataFile}
\alias{getChipType.UnitAnnotationDataFile}
\alias{getPlatform.UnitAnnotationDataFile}
\alias{getUnitAnnotationDataFile}
\alias{getUnitAnnotationDataFile.AromaPlatformInterface}
\alias{getUnitTypes}
\alias{getUnitTypes.UnitTypesFile}
\alias{getUnitTypesFile}
\alias{getUnitTypesFile.AromaPlatform}
\alias{getUnitTypesFile.AromaPlatformInterface}
\alias{nbrOfUnits.UnitAnnotationDataFile}
\alias{nbrOfUnits.UnitTypesFile}
\alias{allocate.AromaUnitTypesFile}
\alias{getChipType.AromaUnitTypesFile}
\alias{getPlatform.AromaUnitTypesFile}
\alias{getUnitTypes.AromaUnitTypesFile}
\alias{importFromUnitTypesFile}
\alias{importFromUnitTypesFile.AromaUnitTypesFile}
\alias{extractRawCopyNumbers.RawSequenceReads}
\alias{yRange}
\alias{yMin}
\alias{yMax}
\alias{yMax.RawGenomicSignals}
\alias{yMin.RawGenomicSignals}
\alias{yRange.RawGenomicSignals}
\alias{byPath.AromaUnitCallSet}
\alias{byPath.AromaUnitGenotypeCallSet}
\alias{setName.RawGenomicSignals}
\alias{getExtensionPattern.AromaCellPositionFile}
\alias{getExtensionPattern.AromaCellSequenceFile}
\alias{getExtensionPattern.AromaUgpFile}
\alias{getExtensionPattern.AromaUnitSignalBinaryFile}
\alias{getExtensionPattern.SampleAnnotationFile}
\alias{getExtensionPattern.TextUnitNamesFile}
\alias{allocate.AromaUflFile}
\alias{AromaUflFile}
\alias{getAromaUflFile}
\alias{getAromaUflFile.UnitAnnotationDataFile}
\alias{getColumnNames.AromaUflFile}
\alias{getExtensionPattern.AromaUflFile}
\alias{getFilenameExtension.AromaUflFile}
\alias{importFromGenericTabularFile.AromaUflFile}
\alias{nbrOfEnzymes}
\alias{nbrOfEnzymes.AromaUflFile}
\alias{readDataFrame.AromaUflFile}
\alias{summaryOfUnits}
\alias{summaryOfUnits.AromaUflFile}
\alias{as.AromaMicroarrayDataSetTuple}
\alias{as.AromaMicroarrayDataSetTuple.AromaMicroarrayDataSet}
\alias{as.AromaMicroarrayDataSetTuple.AromaMicroarrayDataSetTuple}
\alias{as.character.ChromosomalModel}
\alias{as.character.CopyNumberChromosomalModel}
\alias{as.CopyNumberDataFile}
\alias{as.CopyNumberDataFile.CopyNumberDataFile}
\alias{as.CopyNumberDataSet}
\alias{as.CopyNumberDataSet.CopyNumberDataSet}
\alias{as.CopyNumberDataSetTuple}
\alias{as.CopyNumberDataSetTuple.CopyNumberDataSetTuple}
\alias{as.CopyNumberDataSetTuple.list}
\alias{calculateChromosomeStatistics}
\alias{calculateChromosomeStatistics.CopyNumberChromosomalModel}
\alias{clearCache.ChromosomalModel}
\alias{clearCache.CopyNumberChromosomalModel}
\alias{CopyNumberDataFile}
\alias{CopyNumberDataSet}
\alias{CopyNumberDataSetTuple}
\alias{estimateSds}
\alias{extractRawCopyNumbers.CopyNumberChromosomalModel}
\alias{fit}
\alias{fit.ChromosomalModel}
\alias{getAlias.ChromosomalModel}
\alias{getAsteriskTags.ChromosomalModel}
\alias{getAsteriskTags.CopyNumberSegmentationModel}
\alias{getAsteriskTags.HaarSegModel}
\alias{getAverageFile}
\alias{getAverageFile.AromaMicroarrayDataSet}
\alias{getChipEffectFiles}
\alias{getChipEffectFiles.ChromosomalModel}
\alias{getChipTypes.ChromosomalModel}
\alias{getDataFileMatrix}
\alias{getDataFileMatrix.CopyNumberChromosomalModel}
\alias{getFitFunction}
\alias{getFitFunction.CbsModel}
\alias{getFitFunction.CopyNumberSegmentationModel}
\alias{getFitFunction.GladModel}
\alias{getFitFunction.HaarSegModel}
\alias{getFullName.ChromosomalModel}
\alias{getFullNames.ChromosomalModel}
\alias{getGenome}
\alias{getGenome.ChromosomalModel}
\alias{getGenomeData}
\alias{getGenomeData.ChromosomalModel}
\alias{getGenomeFile}
\alias{getGenomeFile.ChromosomalModel}
\alias{getListOfAromaUgpFiles}
\alias{getListOfAromaUgpFiles.ChromosomalModel}
\alias{getListOfChipEffectSets}
\alias{getListOfChipEffectSets.ChromosomalModel}
\alias{getListOfUnitNamesFiles}
\alias{getListOfUnitNamesFiles.ChromosomalModel}
\alias{getListOfUnitTypesFiles}
\alias{getListOfUnitTypesFiles.ChromosomalModel}
\alias{getLog2Ratios}
\alias{getLog2Ratios.CopyNumberSegmentationModel}
\alias{getName.ChromosomalModel}
\alias{getNames.ChromosomalModel}
\alias{getNumberOfFilesAveraged}
\alias{getNumberOfFilesAveraged.CopyNumberDataFile}
\alias{getParentPath}
\alias{getParentPath.ChromosomalModel}
\alias{getPath.ChromosomalModel}
\alias{getRawCnData}
\alias{getRawCnData.CopyNumberChromosomalModel}
\alias{getReferenceSetTuple}
\alias{getReferenceSetTuple.CopyNumberChromosomalModel}
\alias{getRefSetTuple}
\alias{getRefSetTuple.CopyNumberChromosomalModel}
\alias{getRegions}
\alias{getRegions.CopyNumberSegmentationModel}
\alias{getReportPath}
\alias{getReportPath.ChromosomalModel}
\alias{getRootPath.ChromosomalModel}
\alias{getSetTag}
\alias{getSetTag.ChromosomalModel}
\alias{getSetTuple}
\alias{getSetTuple.ChromosomalModel}
\alias{getTags.ChromosomalModel}
\alias{getTags.CopyNumberSegmentationModel}
\alias{getXAM}
\alias{getXAM.AromaMicroarrayDataSet}
\alias{hasAlleleBFractions}
\alias{hasAlleleBFractions.CopyNumberDataFile}
\alias{hasAlleleBFractions.CopyNumberDataSet}
\alias{hasAlleleBFractions.CopyNumberDataSetTuple}
\alias{hasStrandiness}
\alias{hasStrandiness.CopyNumberDataFile}
\alias{hasStrandiness.CopyNumberDataSet}
\alias{hasStrandiness.CopyNumberDataSetTuple}
\alias{isAverageFile}
\alias{isAverageFile.AromaMicroarrayDataFile}
\alias{isPaired}
\alias{isPaired.CopyNumberChromosomalModel}
\alias{onFitAddGenotypeCalls}
\alias{onFitAddGenotypeCalls.default}
\alias{onFitAddGenotypeCalls.GladModel}
\alias{setAlias.ChromosomalModel}
\alias{setGenome}
\alias{setGenome.ChromosomalModel}
\alias{writeRegions}
\alias{writeRegions.CopyNumberSegmentationModel}
\alias{writeRegions.GladModel}
\alias{getAsteriskTags.RawCopyNumberModel}
\alias{getSetTag.RawCopyNumberModel}
\alias{AromaUnitTotalCnBinarySetTuple}
\alias{as.AromaUnitTotalCnBinarySetTuple}
\alias{as.AromaUnitTotalCnBinarySetTuple.AromaUnitTotalCnBinarySet}
\alias{as.AromaUnitTotalCnBinarySetTuple.AromaUnitTotalCnBinarySetTuple}
\alias{as.CopyNumberDataSetTuple.AromaUnitTotalCnBinarySet}
\alias{getAM}
\alias{getAM.AromaUnitTotalCnBinaryFile}
\alias{getListOfUnitNamesFiles.AromaUnitTotalCnBinarySetTuple}
\alias{getUnitNamesFile.AromaUnitTotalCnBinarySet}
\alias{getXAM.AromaMicroarrayDataFile}
\alias{getXAM.AromaUnitTotalCnBinaryFile}
\alias{hasAlleleBFractions.AromaUnitTotalCnBinaryFile}
\alias{hasStrandiness.AromaUnitTotalCnBinaryFile}
\alias{isAverageFile.AromaUnitSignalBinaryFile}
\alias{nbrOfArrays.AromaUnitTotalCnBinarySet}
\alias{getAverageFile.AromaUnitTotalCnBinarySet}
\alias{getNumberOfFilesAveraged.AromaUnitTotalCnBinaryFile}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
