%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{[.AromaTabularBinaryFile}
\alias{[[.AromaTabularBinaryFile}
\alias{[<-.AromaTabularBinaryFile}
\alias{addBy}
\alias{addBy.RawGenomicSignals}
\alias{addIncludes}
\alias{addIncludes.Explorer}
\alias{addIndexFile}
\alias{addIndexFile.ChromosomeExplorer}
\alias{addIndexFile.Explorer}
\alias{addLocusFields}
\alias{addLocusFields.RawGenomicSignals}
\alias{allocate}
\alias{allocate.AromaCellPositionFile}
\alias{allocate.AromaCellSequenceFile}
\alias{allocate.AromaMicroarrayTabularBinaryFile}
\alias{allocate.AromaUflFile}
\alias{allocate.AromaUgpFile}
\alias{allocate.AromaUnitCallFile}
\alias{allocate.AromaUnitChromosomeTabularBinaryFile}
\alias{allocate.AromaUnitGenotypeCallFile}
\alias{allocate.AromaUnitSignalBinaryFile}
\alias{allocate.AromaUnitTypesFile}
\alias{allocateFromUnitAnnotationDataFile}
\alias{allocateFromUnitAnnotationDataFile.AromaUnitSignalBinaryFile}
\alias{allocateFromUnitAnnotationDataFile.AromaUnitTabularBinaryFile}
\alias{allocateFromUnitNamesFile}
\alias{allocateFromUnitNamesFile.AromaUnitSignalBinaryFile}
\alias{allocateFromUnitNamesFile.AromaUnitTabularBinaryFile}
\alias{append.RawGenomicSignals}
\alias{apply}
\alias{apply.default}
\alias{apply.SampleAnnotationFile}
\alias{applyBinaryOperator}
\alias{applyBinaryOperator.RawGenomicSignals}
\alias{applyRegions}
\alias{applyRegions.CopyNumberOutliers}
\alias{applyRows}
\alias{applyRows.CopyNumberRegions}
\alias{AromaCellSequenceFile}
\alias{AromaCore}
\alias{AromaPackage}
\alias{AromaPlatformInterface}
\alias{aromaSettings}
\alias{AromaSettings}
\alias{AromaUflFile}
\alias{AromaUgpFile}
\alias{AromaUnitChromosomeTabularBinaryFile}
\alias{AromaUnitTotalCnBinaryFileList}
\alias{AromaUnitTotalCnBinarySetTuple}
\alias{as.AromaMicroarrayDataSetList}
\alias{as.AromaMicroarrayDataSetList.AromaMicroarrayDataSet}
\alias{as.AromaMicroarrayDataSetTuple}
\alias{as.AromaMicroarrayDataSetTuple.AromaMicroarrayDataSet}
\alias{as.AromaMicroarrayDataSetTuple.AromaMicroarrayDataSetTuple}
\alias{as.AromaUnitTotalCnBinarySetTuple}
\alias{as.AromaUnitTotalCnBinarySetTuple.AromaUnitTotalCnBinarySet}
\alias{as.AromaUnitTotalCnBinarySetTuple.AromaUnitTotalCnBinarySetTuple}
\alias{as.character.AromaMicroarrayDataSetTuple}
\alias{as.character.AromaMicroarrayTabularBinaryFile}
\alias{as.character.AromaTabularBinaryFile}
\alias{as.character.AromaTransform}
\alias{as.character.AromaUnitSignalBinaryFile}
\alias{as.character.ChromosomalModel}
\alias{as.character.ChromosomeExplorer}
\alias{as.character.CopyNumberChromosomalModel}
\alias{as.character.CopyNumberOutliers}
\alias{as.character.CopyNumberRegions}
\alias{as.character.Explorer}
\alias{as.character.RawGenomicSignals}
\alias{as.character.TextUnitNamesFile}
\alias{as.CopyNumberDataFile}
\alias{as.CopyNumberDataFile.CopyNumberDataFile}
\alias{as.CopyNumberDataSet}
\alias{as.CopyNumberDataSet.CopyNumberDataSet}
\alias{as.CopyNumberDataSetTuple}
\alias{as.CopyNumberDataSetTuple.AromaUnitTotalCnBinarySet}
\alias{as.CopyNumberDataSetTuple.CopyNumberDataSetTuple}
\alias{as.CopyNumberDataSetTuple.list}
\alias{as.data.frame.CopyNumberOutliers}
\alias{as.data.frame.CopyNumberRegions}
\alias{as.data.frame.RawCopyNumbers}
\alias{as.data.frame.RawGenomicSignals}
\alias{as.data.frame.SegmentedGenomicSignalsInterface}
\alias{as.GrayscaleImage}
\alias{as.TrueColorImage}
\alias{as.TrueColorImage.Image}
\alias{as.TrueColorImage.matrix}
\alias{asMatrixOfFiles}
\alias{asMatrixOfFiles.AromaMicroarrayDataSetTuple}
\alias{asThis}
\alias{asThis.BasicObject}
\alias{asThis.Object}
\alias{backtransformGenotypeCone}
\alias{backtransformGenotypeCone.matrix}
\alias{backtransformMultiDimensionalCone}
\alias{backtransformMultiDimensionalCone.matrix}
\alias{barSequence}
\alias{barSequence.ProbePositionEffects}
\alias{binnedSmoothing.RawGenomicSignals}
\alias{binnedSmoothingByState}
\alias{binnedSmoothingByState.SegmentedGenomicSignalsInterface}
\alias{binnedSums}
\alias{binnedSums.RawSequenceReads}
\alias{binScatter}
\alias{binScatter.matrix}
\alias{byChipType}
\alias{byChipType.AromaCellPositionFile}
\alias{byChipType.AromaMicroarrayTabularBinaryFile}
\alias{byChipType.AromaUnitTabularBinaryFile}
\alias{byChipType.UnitAnnotationDataFile}
\alias{byName.AromaPlatform}
\alias{byName.AromaUnitFracBCnBinarySet}
\alias{byName.AromaUnitGenotypeCallSet}
\alias{byName.AromaUnitSignalBinarySet}
\alias{byName.AromaUnitTotalCnBinarySet}
\alias{byPath.AromaMicroarrayDataSetTuple}
\alias{byPath.AromaUnitCallSet}
\alias{byPath.AromaUnitGenotypeCallSet}
\alias{calcMargins}
\alias{calcMargins.matrix}
\alias{calculateAverageColumnAcrossFiles}
\alias{calculateAverageColumnAcrossFiles.GenericTabularFileSet}
\alias{calculateChromosomeStatistics}
\alias{calculateChromosomeStatistics.CopyNumberChromosomalModel}
\alias{calculateRatios}
\alias{calculateRatios.CopyNumberChromosomalModel}
\alias{cnRange}
\alias{cnRange.RawCopyNumbers}
\alias{colGaussianSmoothing}
\alias{colGaussianSmoothing.matrix}
\alias{colMeans}
\alias{colMeans.AromaTabularBinaryFile}
\alias{colMeans.default}
\alias{colMedians.AromaTabularBinaryFile}
\alias{colorize}
\alias{colorize.Image}
\alias{colStats}
\alias{colStats.AromaTabularBinaryFile}
\alias{colSums}
\alias{colSums.AromaTabularBinaryFile}
\alias{colSums.default}
\alias{convertTable}
\alias{convertTable.default}
\alias{CopyNumberDataFile}
\alias{CopyNumberDataSet}
\alias{CopyNumberDataSetTuple}
\alias{CopyNumberOutliers}
\alias{CopyNumberRegions}
\alias{countBases}
\alias{countBases.AromaCellSequenceFile}
\alias{countBasesInternal}
\alias{countBasesInternal.AromaCellSequenceFile}
\alias{dimnames<-.AromaTabularBinaryFile}
\alias{display}
\alias{display.ChromosomeExplorer}
\alias{display.Image}
\alias{divideBy}
\alias{divideBy.RawGenomicSignals}
\alias{drawCnRegions}
\alias{drawCnRegions.DNAcopy}
\alias{drawCnRegions.HaarSeg}
\alias{drawCnRegions.profileCGH}
\alias{drawCytoband}
\alias{drawCytoband.ChromosomalModel}
\alias{drawCytoband.profileCGH}
\alias{drawCytoband2}
\alias{drawCytoband2.default}
\alias{drawExtraAnnotations}
\alias{drawExtraAnnotations.default}
\alias{drawExtraAnnotations.profileCGH}
\alias{drawLevels}
\alias{drawLevels.CopyNumberOutliers}
\alias{drawLevels.CopyNumberRegions}
\alias{estimateSds}
\alias{estimateStandardDeviation}
\alias{exportFracBDiffSet}
\alias{exportFracBDiffSet.AromaUnitFracBCnBinarySet}
\alias{exportTotalCnRatioSet}
\alias{exportTotalCnRatioSet.AromaUnitTotalCnBinarySet}
\alias{extractByChromosome}
\alias{extractByChromosome.AromaUnitChromosomeTabularBinaryFile}
\alias{extractCallArray}
\alias{extractCallArray.AromaUnitCallFile}
\alias{extractCallArray.AromaUnitCallSet}
\alias{extractCalls}
\alias{extractCalls.AromaUnitCallFile}
\alias{extractCalls.AromaUnitCallSet}
\alias{extractCNRs}
\alias{extractCNRs.default}
\alias{extractCopyNumberOutliers}
\alias{extractCopyNumberOutliers.DNAcopy}
\alias{extractCopyNumberOutliers.profileCGH}
\alias{extractCopyNumberRegions}
\alias{extractCopyNumberRegions.default}
\alias{extractCopyNumberRegions.DNAcopy}
\alias{extractCopyNumberRegions.HaarSeg}
\alias{extractCopyNumberRegions.profileCGH}
\alias{extractDataForSegmentation}
\alias{extractDataForSegmentation.RawGenomicSignals}
\alias{extractGenotypeMatrix}
\alias{extractGenotypeMatrix.AromaUnitCallSet}
\alias{extractGenotypeMatrix.AromaUnitGenotypeCallFile}
\alias{extractGenotypes}
\alias{extractGenotypes.AromaUnitGenotypeCallFile}
\alias{extractGenotypes.AromaUnitGenotypeCallSet}
\alias{extractIGV}
\alias{extractIGV.CopyNumberRegions}
\alias{extractMatrix.AromaUnitCallFile}
\alias{extractMatrix.AromaUnitSignalBinaryFile}
\alias{extractMergedRawCopyNumbers}
\alias{extractMergedRawCopyNumbers.AromaUnitTotalCnBinaryFileList}
\alias{extractOld}
\alias{extractOld.AromaMicroarrayDataSetTuple}
\alias{extractRawAlleleBFractions}
\alias{extractRawAlleleBFractions.AromaUnitFracBCnBinaryFile}
\alias{extractRawAlleleBFractions.default}
\alias{extractRawCNs}
\alias{extractRawCNs.default}
\alias{extractRawCopyNumbers}
\alias{extractRawCopyNumbers.AromaUnitTotalCnBinaryFile}
\alias{extractRawCopyNumbers.AromaUnitTotalCnBinaryFileList}
\alias{extractRawCopyNumbers.default}
\alias{extractRawCopyNumbers.DNAcopy}
\alias{extractRawCopyNumbers.HaarSeg}
\alias{extractRawCopyNumbers.profileCGH}
\alias{extractRawCopyNumbers.RawCopyNumbers}
\alias{extractRawCopyNumbers.RawSequenceReads}
\alias{extractRawGenomicSignals}
\alias{extractRawGenomicSignals.AromaUnitSignalBinaryFile}
\alias{extractRawGenomicSignals.AromaUnitTotalCnBinaryFileList}
\alias{extractRawGenomicSignals.default}
\alias{extractRawMirroredAlleleBFractions}
\alias{extractRawMirroredAlleleBFractions.default}
\alias{extractRawMirroredAlleleBFractions.RawAlleleBFractions}
\alias{extractRegion}
\alias{extractRegion.RawGenomicSignals}
\alias{extractSubset}
\alias{extractSubset.RawGenomicSignals}
\alias{extractSubsetByState}
\alias{extractSubsetByState.SegmentedGenomicSignalsInterface}
\alias{findByChipType}
\alias{findByChipType.AromaMicroarrayTabularBinaryFile}
\alias{findByChipType.TextUnitNamesFile}
\alias{findByName.AromaUnitCallSet}
\alias{findByName.AromaUnitSignalBinarySet}
\alias{findPngDevice}
\alias{findPngDevice.default}
\alias{findSAFs}
\alias{findSAFs.SampleAnnotationSet}
\alias{findUnitNamesFile}
\alias{findUnitNamesFile.AromaPlatform}
\alias{findUnitsTodo}
\alias{findUnitsTodo.AromaUnitCallFile}
\alias{findUnitsTodo.AromaUnitCallSet}
\alias{findUnitTypesFile}
\alias{findUnitTypesFile.AromaPlatform}
\alias{fit}
\alias{fit.ChromosomalModel}
\alias{fit2d}
\alias{fit2d.matrix}
\alias{fitGenotypeCone}
\alias{fitGenotypeConeByExpectile}
\alias{fitGenotypeConeByExpectile.matrix}
\alias{fitGenotypeConeBySfit}
\alias{fitMultiDimensionalCone}
\alias{fitProbePositionEffects}
\alias{fitProbePositionEffects.numeric}
\alias{fitSplineBlockPolish}
\alias{fitSplineBlockPolish.matrix}
\alias{fitWHLAPLM}
\alias{fitWHLAPLM.matrix}
\alias{fitWHRCModel}
\alias{fitWHRCModel.matrix}
\alias{fitWLAPLM}
\alias{fitWLAPLM.matrix}
\alias{fitWRCModel}
\alias{fitWRCModel.matrix}
\alias{fitWRMA}
\alias{fitWRMA.matrix}
\alias{fixSearchPath}
\alias{fixSearchPath.AromaCore}
\alias{fixSearchPath.AromaPackage}
\alias{fixSearchPathInternal}
\alias{fixSearchPathInternal.AromaPackage}
\alias{fromFile.AromaUnitSignalBinaryFile}
\alias{fromPath}
\alias{fromPath.SampleAnnotationFile}
\alias{fromPath.SampleAnnotationSet}
\alias{gaussianSmoothing}
\alias{gaussianSmoothing.matrix}
\alias{gaussianSmoothing.numeric}
\alias{gaussianSmoothing.RawGenomicSignals}
\alias{gaussKernel}
\alias{getAlias.ChromosomalModel}
\alias{getAM}
\alias{getAM.AromaUnitTotalCnBinaryFile}
\alias{getAromaPlatform}
\alias{getAromaPlatform.AromaPlatformInterface}
\alias{getAromaUflFile}
\alias{getAromaUflFile.UnitAnnotationDataFile}
\alias{getAromaUgpFile}
\alias{getAromaUgpFile.AromaPlatform}
\alias{getAromaUgpFile.AromaUnitSignalBinaryFile}
\alias{getAromaUgpFile.AromaUnitSignalBinarySet}
\alias{getAromaUgpFile.UnitAnnotationDataFile}
\alias{getArrays}
\alias{getArrays.AromaMicroarrayDataSetTuple}
\alias{getArrays.ChromosomalModel}
\alias{getArrays.Explorer}
\alias{getArraysOfInput}
\alias{getArraysOfInput.ChromosomeExplorer}
\alias{getArraysOfInput.Explorer}
\alias{getArrayTuple}
\alias{getArrayTuple.AromaMicroarrayDataSetTuple}
\alias{getAsteriskTags.AromaMicroarrayDataSetTuple}
\alias{getAsteriskTags.AromaTransform}
\alias{getAsteriskTags.ChromosomalModel}
\alias{getAsteriskTags.CopyNumberSegmentationModel}
\alias{getAsteriskTags.Explorer}
\alias{getAsteriskTags.HaarSegModel}
\alias{getAsteriskTags.RawCopyNumberModel}
\alias{getAttributeXY}
\alias{getAttributeXY.AromaMicroarrayDataFile}
\alias{getAverageFile}
\alias{getAverageFile.AromaMicroarrayDataSet}
\alias{getAverageFile.AromaUnitTotalCnBinarySet}
\alias{getBytesPerColumn}
\alias{getBytesPerColumn.AromaTabularBinaryFile}
\alias{getChipEffectFiles}
\alias{getChipEffectFiles.ChromosomalModel}
\alias{getChipType}
\alias{getChipType.AromaMicroarrayDataFile}
\alias{getChipType.AromaMicroarrayDataSet}
\alias{getChipType.AromaMicroarrayTabularBinaryFile}
\alias{getChipType.AromaUnitSignalBinaryFile}
\alias{getChipType.AromaUnitSignalBinarySet}
\alias{getChipType.AromaUnitTypesFile}
\alias{getChipType.TextUnitNamesFile}
\alias{getChipType.UnitAnnotationDataFile}
\alias{getChipTypes}
\alias{getChipTypes.AromaMicroarrayDataSetTuple}
\alias{getChipTypes.ChromosomalModel}
\alias{getChromosome}
\alias{getChromosome.Arguments}
\alias{getChromosome.RawGenomicSignals}
\alias{getChromosomes}
\alias{getChromosomes.Arguments}
\alias{getChromosomes.AromaUnitChromosomeTabularBinaryFile}
\alias{getCn}
\alias{getCn.RawCopyNumbers}
\alias{getCNs}
\alias{getCNs.RawCopyNumbers}
\alias{getColClasses}
\alias{getColClasses.AromaTabularBinaryFile}
\alias{getColumnNames.AromaCellPositionFile}
\alias{getColumnNames.AromaCellSequenceFile}
\alias{getColumnNames.AromaTabularBinaryFile}
\alias{getColumnNames.AromaUflFile}
\alias{getColumnNames.AromaUgpFile}
\alias{getColumnNames.AromaUnitChromosomeTabularBinaryFile}
\alias{getCommonListElements}
\alias{getDataFileMatrix}
\alias{getDataFileMatrix.CopyNumberChromosomalModel}
\alias{getDeviceResolution}
\alias{getDeviceResolution.default}
\alias{getEffects}
\alias{getEffects.ProbePositionEffects}
\alias{getExpectedOutputFiles}
\alias{getExpectedOutputFiles.AromaTransform}
\alias{getExpectedOutputFullnames}
\alias{getExpectedOutputFullnames.AromaTransform}
\alias{getExtensionPattern.AromaCellPositionFile}
\alias{getExtensionPattern.AromaCellSequenceFile}
\alias{getExtensionPattern.AromaUflFile}
\alias{getExtensionPattern.AromaUgpFile}
\alias{getExtensionPattern.AromaUnitSignalBinaryFile}
\alias{getExtensionPattern.SampleAnnotationFile}
\alias{getExtensionPattern.TextUnitNamesFile}
\alias{getFilenameExtension.AromaCellPositionFile}
\alias{getFilenameExtension.AromaCellSequenceFile}
\alias{getFilenameExtension.AromaMicroarrayTabularBinaryFile}
\alias{getFilenameExtension.AromaUflFile}
\alias{getFilenameExtension.AromaUgpFile}
\alias{getFilenameExtension.AromaUnitChromosomeTabularBinaryFile}
\alias{getFilenameExtension.AromaUnitSignalBinaryFile}
\alias{getFilenameExtension.TextUnitNamesFile}
\alias{getFitFunction}
\alias{getFitFunction.CbsModel}
\alias{getFitFunction.CopyNumberSegmentationModel}
\alias{getFitFunction.GladModel}
\alias{getFitFunction.HaarSegModel}
\alias{getFullName.ChromosomalModel}
\alias{getFullName.Explorer}
\alias{getFullNames.AromaMicroarrayDataSetTuple}
\alias{getFullNames.ChromosomalModel}
\alias{getFullNames.ChromosomeExplorer}
\alias{getFullNames.CopyNumberSegmentationModel}
\alias{getGenome}
\alias{getGenome.ChromosomalModel}
\alias{getGenomeData}
\alias{getGenomeData.ChromosomalModel}
\alias{getGenomeFile}
\alias{getGenomeFile.ChromosomalModel}
\alias{getGenomeVersion}
\alias{getGenomeVersion.AromaUnitChromosomeTabularBinaryFile}
\alias{getHeaderParameters}
\alias{getHeaderParameters.TextUnitNamesFile}
\alias{getImage}
\alias{getImage.matrix}
\alias{getIncludePath}
\alias{getIncludePath.Explorer}
\alias{getInputDataSet}
\alias{getListOfAromaUgpFiles}
\alias{getListOfAromaUgpFiles.ChromosomalModel}
\alias{getListOfChipEffectSets}
\alias{getListOfChipEffectSets.ChromosomalModel}
\alias{getListOfUnitNamesFiles}
\alias{getListOfUnitNamesFiles.AromaUnitTotalCnBinarySetTuple}
\alias{getListOfUnitNamesFiles.ChromosomalModel}
\alias{getListOfUnitTypesFiles}
\alias{getListOfUnitTypesFiles.ChromosomalModel}
\alias{getLociFields}
\alias{getLociFields.RawGenomicSignals}
\alias{getLocusFields}
\alias{getLocusFields.RawGenomicSignals}
\alias{getLog2Ratios}
\alias{getLog2Ratios.CopyNumberSegmentationModel}
\alias{getMainPath}
\alias{getMainPath.Explorer}
\alias{getMaxLengthRepeats}
\alias{getMaxLengthRepeats.AromaCellSequenceFile}
\alias{getModel}
\alias{getName.AromaPlatform}
\alias{getName.ChromosomalModel}
\alias{getName.RawGenomicSignals}
\alias{getNameOfInput}
\alias{getNameOfInput.ChromosomeExplorer}
\alias{getNameOfInput.Explorer}
\alias{getNames.ChromosomeExplorer}
\alias{getNames.CopyNumberChromosomalModel}
\alias{getNumberOfFilesAveraged}
\alias{getNumberOfFilesAveraged.CopyNumberDataFile}
\alias{getOutputDataSet}
\alias{getOutputDataSet0}
\alias{getOutputDataSet0.AromaTransform}
\alias{getOutputFiles}
\alias{getOutputFiles.AromaTransform}
\alias{getPairedNames}
\alias{getPairedNames.CopyNumberChromosomalModel}
\alias{getParallelSafe}
\alias{getParallelSafe.Explorer}
\alias{getParametersAsString}
\alias{getParentPath}
\alias{getParentPath.ChromosomalModel}
\alias{getPath.ChromosomalModel}
\alias{getPath.ChromosomeExplorer}
\alias{getPatterns}
\alias{getPatterns.SampleAnnotationFile}
\alias{getPhysicalPositions}
\alias{getPhysicalPositions.RawCopyNumbers}
\alias{getPlatform}
\alias{getPlatform.AromaMicroarrayDataFile}
\alias{getPlatform.AromaMicroarrayDataSet}
\alias{getPlatform.AromaMicroarrayTabularBinaryFile}
\alias{getPlatform.AromaPlatformInterface}
\alias{getPlatform.AromaUnitSignalBinaryFile}
\alias{getPlatform.AromaUnitSignalBinarySet}
\alias{getPlatform.AromaUnitTypesFile}
\alias{getPlatform.TextUnitNamesFile}
\alias{getPlatform.UnitAnnotationDataFile}
\alias{getPloidy}
\alias{getPloidy.AromaMicroarrayDataFile}
\alias{getPositions}
\alias{getPositions.AromaUgpFile}
\alias{getPositions.RawGenomicSignals}
\alias{getProbeLength}
\alias{getProbeLength.AromaCellSequenceFile}
\alias{getProbePositionEffectDesignMatrix}
\alias{getProbePositionEffectDesignMatrix.character}
\alias{getProbePositionEffectDesignMatrix.raw}
\alias{getRam}
\alias{getRam.AromaSettings}
\alias{getRawCnData}
\alias{getRawCnData.CopyNumberChromosomalModel}
\alias{getReferenceSetTuple}
\alias{getReferenceSetTuple.CopyNumberChromosomalModel}
\alias{getRefSetTuple}
\alias{getRefSetTuple.CopyNumberChromosomalModel}
\alias{getRegions}
\alias{getRegions.CopyNumberSegmentationModel}
\alias{getReportPath}
\alias{getReportPath.ChromosomalModel}
\alias{getReportPathPattern}
\alias{getReportPathPattern.Explorer}
\alias{getRootPath}
\alias{getRootPath.ChromosomalModel}
\alias{getSampleLabels}
\alias{getSampleLabels.ChromosomeExplorer}
\alias{getSampleLayerName}
\alias{getSampleLayerName.Explorer}
\alias{getSampleLayerPrefix}
\alias{getSampleLayerPrefix.Explorer}
\alias{getSets.AromaMicroarrayDataSetTuple}
\alias{getSets.ChromosomalModel}
\alias{getSetTag}
\alias{getSetTag.ChromosomalModel}
\alias{getSetTag.RawCopyNumberModel}
\alias{getSetTuple}
\alias{getSetTuple.ChromosomalModel}
\alias{getSigma}
\alias{getSigma.RawGenomicSignals}
\alias{getSignals}
\alias{getSignals.RawGenomicSignals}
\alias{getSnpNucleotides}
\alias{getSnpNucleotides.AromaCellSequenceFile}
\alias{getSnpPositions}
\alias{getSnpPositions.AromaCellSequenceFile}
\alias{getSnpShifts}
\alias{getSnpShifts.AromaCellSequenceFile}
\alias{getStateColorMap}
\alias{getStateColorMap.SegmentedGenomicSignalsInterface}
\alias{getStateColors}
\alias{getStateColors.SegmentedGenomicSignalsInterface}
\alias{getStates}
\alias{getStates.SegmentedGenomicSignalsInterface}
\alias{getSubname}
\alias{getSubname.Explorer}
\alias{getTableOfArrays}
\alias{getTableOfArrays.AromaMicroarrayDataSetTuple}
\alias{getTableOfArrays.ChromosomalModel}
\alias{getTags.AromaMicroarrayDataSetTuple}
\alias{getTags.ChromosomalModel}
\alias{getTags.CopyNumberSegmentationModel}
\alias{getTagsOfInput}
\alias{getTagsOfInput.ChromosomeExplorer}
\alias{getTagsOfInput.Explorer}
\alias{getTemplatePath}
\alias{getTemplatePath.Explorer}
\alias{getUniqueStates}
\alias{getUniqueStates.SegmentedGenomicSignalsInterface}
\alias{getUnitAnnotationDataFile}
\alias{getUnitAnnotationDataFile.AromaPlatformInterface}
\alias{getUnitNames}
\alias{getUnitNames.TextUnitNamesFile}
\alias{getUnitNames.UnitNamesFile}
\alias{getUnitNamesFile}
\alias{getUnitNamesFile.AromaPlatform}
\alias{getUnitNamesFile.AromaPlatformInterface}
\alias{getUnitNamesFile.AromaUnitTotalCnBinarySet}
\alias{getUnitsAt}
\alias{getUnitsAt.AromaUgpFile}
\alias{getUnitsOnChromosome}
\alias{getUnitsOnChromosome.AromaUgpFile}
\alias{getUnitsOnChromosome.AromaUnitChromosomeTabularBinaryFile}
\alias{getUnitsOnChromosomes}
\alias{getUnitsOnChromosomes.AromaUnitChromosomeTabularBinaryFile}
\alias{getUnitTypes}
\alias{getUnitTypes.AromaUnitTypesFile}
\alias{getUnitTypes.UnitTypesFile}
\alias{getUnitTypesFile}
\alias{getUnitTypesFile.AromaPlatform}
\alias{getUnitTypesFile.AromaPlatformInterface}
\alias{getVerbose.AromaSettings}
\alias{getWeights}
\alias{getWeights.RawGenomicSignals}
\alias{getXAM}
\alias{getXAM.AromaMicroarrayDataFile}
\alias{getXAM.AromaUnitTotalCnBinaryFile}
\alias{getXScale}
\alias{getXScale.RawGenomicSignals}
\alias{getXY}
\alias{getXY.RawGenomicSignals}
\alias{getYScale}
\alias{getYScale.RawGenomicSignals}
\alias{getZooms}
\alias{getZooms.ChromosomeExplorer}
\alias{groupBySnpNucleotides}
\alias{groupBySnpNucleotides.AromaCellSequenceFile}
\alias{hasAlleleBFractions}
\alias{hasAlleleBFractions.AromaUnitTotalCnBinaryFile}
\alias{hasAlleleBFractions.CopyNumberDataFile}
\alias{hasAlleleBFractions.CopyNumberDataSet}
\alias{hasAlleleBFractions.CopyNumberDataSetTuple}
\alias{hasAttributeXY}
\alias{hasAttributeXY.AromaMicroarrayDataFile}
\alias{hasStrandiness}
\alias{hasStrandiness.AromaUnitTotalCnBinaryFile}
\alias{hasStrandiness.CopyNumberDataFile}
\alias{hasStrandiness.CopyNumberDataSet}
\alias{hasStrandiness.CopyNumberDataSetTuple}
\alias{hasWeights}
\alias{hasWeights.RawGenomicSignals}
\alias{importFrom}
\alias{importFrom.AromaTabularBinaryFile}
\alias{importFromGenericTabularFile}
\alias{importFromGenericTabularFile.AromaUflFile}
\alias{importFromGenericTabularFile.AromaUgpFile}
\alias{importFromTable}
\alias{importFromTable.FileMatrix}
\alias{importFromUnitTypesFile}
\alias{importFromUnitTypesFile.AromaUnitTypesFile}
\alias{indexOf.AromaMicroarrayDataSetTuple}
\alias{indexOf.ChromosomalModel}
\alias{indexOf.ChromosomeExplorer}
\alias{indexOfColumn}
\alias{indexOfColumn.AromaUnitChromosomeTabularBinaryFile}
\alias{indexOfUnits}
\alias{indexOfUnits.AromaUnitTabularBinaryFile}
\alias{interleave}
\alias{interleave.Image}
\alias{isAverageFile}
\alias{isAverageFile.AromaMicroarrayDataFile}
\alias{isAverageFile.AromaUnitSignalBinaryFile}
\alias{isHeterozygote}
\alias{isHeterozygote.AromaUnitGenotypeCallFile}
\alias{isHeterozygous}
\alias{isHeterozygous.AromaUnitGenotypeCallFile}
\alias{isHomozygote}
\alias{isHomozygote.AromaUnitGenotypeCallFile}
\alias{isHomozygous}
\alias{isHomozygous.AromaUnitGenotypeCallFile}
\alias{isMissing}
\alias{isMissing.AromaCellPositionFile}
\alias{isMissing.AromaCellSequenceFile}
\alias{isPaired}
\alias{isPaired.CopyNumberChromosomalModel}
\alias{kernelSmoothing.RawGenomicSignals}
\alias{kernelSmoothingByState}
\alias{kernelSmoothingByState.SegmentedGenomicSignalsInterface}
\alias{lapply.AromaTabularBinaryFile}
\alias{lapplyInChunks}
\alias{lapplyInChunks.numeric}
\alias{library}
\alias{lines.CopyNumberOutliers}
\alias{lines.CopyNumberRegions}
\alias{lines.RawGenomicSignals}
\alias{listToXml}
\alias{listToXml.list}
\alias{locallyUnique}
\alias{locallyUnique.default}
\alias{log2abs}
\alias{log2center}
\alias{log2neg}
\alias{log2pos}
\alias{matchPatterns}
\alias{matchPatterns.SampleAnnotationFile}
\alias{matrixBlockPolish}
\alias{mergeBoxplotStats}
\alias{multiplyBy}
\alias{multiplyBy.RawGenomicSignals}
\alias{nbrOfArrays}
\alias{nbrOfArrays.AromaMicroarrayDataSet}
\alias{nbrOfArrays.AromaMicroarrayDataSetTuple}
\alias{nbrOfArrays.AromaUnitTotalCnBinarySet}
\alias{nbrOfCells}
\alias{nbrOfCells.AromaCellTabularBinaryFile}
\alias{nbrOfChipTypes}
\alias{nbrOfColumns.AromaTabularBinaryFile}
\alias{nbrOfEnzymes}
\alias{nbrOfEnzymes.AromaUflFile}
\alias{nbrOfLoci}
\alias{nbrOfLoci.RawGenomicSignals}
\alias{nbrOfReads}
\alias{nbrOfReads.RawSequenceReads}
\alias{nbrOfRegions}
\alias{nbrOfRegions.CopyNumberOutliers}
\alias{nbrOfRegions.CopyNumberRegions}
\alias{nbrOfRows.AromaTabularBinaryFile}
\alias{nbrOfUnits}
\alias{nbrOfUnits.AromaUnitSignalBinaryFile}
\alias{nbrOfUnits.AromaUnitTabularBinaryFile}
\alias{nbrOfUnits.TextUnitNamesFile}
\alias{nbrOfUnits.UnitAnnotationDataFile}
\alias{nbrOfUnits.UnitNamesFile}
\alias{nbrOfUnits.UnitTypesFile}
\alias{newPlot}
\alias{newPlot.CopyNumberChromosomalModel}
\alias{norm2d}
\alias{norm2d.matrix}
\alias{normalizeGenotypeCone}
\alias{normalizeGenotypeCone.matrix}
\alias{onFitAddGenotypeCalls}
\alias{onFitAddGenotypeCalls.default}
\alias{onFitAddGenotypeCalls.GladModel}
\alias{patch}
\alias{patch.AromaPackage}
\alias{plot.BinnedScatter}
\alias{plot.CopyNumberSegmentationModel}
\alias{plot.ProbePositionEffects}
\alias{plot.profileCGH}
\alias{plot.RawAlleleBFractions}
\alias{plot.RawCopyNumbers}
\alias{plot.RawGenomicSignals}
\alias{plot.RawMirroredAlleleBFractions}
\alias{plot.RawSequenceReads}
\alias{plot.SegmentedGenomicSignalsInterface}
\alias{plotAxesLayers}
\alias{plotAxesLayers.CopyNumberChromosomalModel}
\alias{plotChromosomesLayers}
\alias{plotChromosomesLayers.CopyNumberChromosomalModel}
\alias{plotCopyNumberRegionLayers}
\alias{plotCopyNumberRegionLayers.CopyNumberSegmentationModel}
\alias{plotCytobandLayers}
\alias{plotCytobandLayers.CopyNumberChromosomalModel}
\alias{plotFitLayers}
\alias{plotFitLayers.CopyNumberChromosomalModel}
\alias{plotGridHorizontalLayers}
\alias{plotGridHorizontalLayers.CopyNumberChromosomalModel}
\alias{plotProfile2}
\alias{plotProfile2.profileCGH}
\alias{plotRawCNs}
\alias{plotRawCNs.profileCGH}
\alias{plotRawCopyNumbers}
\alias{plotRawCopyNumbers.CopyNumberChromosomalModel}
\alias{plotSampleLayers}
\alias{plotSampleLayers.CopyNumberChromosomalModel}
\alias{points.BinnedScatter}
\alias{points.RawGenomicSignals}
\alias{points.SegmentedGenomicSignalsInterface}
\alias{pointsRawCNs}
\alias{pointsRawCNs.default}
\alias{pointsRawCNs.profileCGH}
\alias{pointsSequence}
\alias{pointsSequence.ProbePositionEffects}
\alias{predict.ProbePositionEffects}
\alias{process}
\alias{readColumns.AromaTabularBinaryFile}
\alias{readData}
\alias{readData.SampleAnnotationFile}
\alias{readDataFrame.AromaTabularBinaryFile}
\alias{readDataFrame.AromaUflFile}
\alias{readDataFrame.AromaUnitChromosomeTabularBinaryFile}
\alias{readDataFrame.AromaUnitSignalBinaryFile}
\alias{readDataFrame.SampleAnnotationFile}
\alias{readFooter}
\alias{readHeader}
\alias{readHeader.AromaTabularBinaryFile}
\alias{readNeighborSequenceMatrix}
\alias{readNeighborSequenceMatrix.AromaCellSequenceFile}
\alias{readPairSequenceMatrix}
\alias{readPairSequenceMatrix.AromaCellSequenceFile}
\alias{readPositions}
\alias{readPositions.AromaCellPositionFile}
\alias{readRawFooter}
\alias{readRawFooter.AromaTabularBinaryFile}
\alias{readSequenceMatrix}
\alias{readSequenceMatrix.AromaCellSequenceFile}
\alias{readSequences}
\alias{readSequences.AromaCellSequenceFile}
\alias{readTargetStrands}
\alias{readTargetStrands.AromaCellSequenceFile}
\alias{remap}
\alias{remap.default}
\alias{reorder.BinnedScatter}
\alias{require}
\alias{rescale}
\alias{rescale.Image}
\alias{rgbTransform}
\alias{rgbTransform.Image}
\alias{SampleAnnotationFile}
\alias{SampleAnnotationSet}
\alias{segmentByCBS}
\alias{segmentByGLAD}
\alias{segmentByHaarSeg}
\alias{setAlias.ChromosomalModel}
\alias{setArrays}
\alias{setAttributesBy.AromaMicroarrayDataSet}
\alias{setAttributesBySampleAnnotationFile}
\alias{setAttributesBySampleAnnotationFile.AromaMicroarrayDataSet}
\alias{setAttributesBySampleAnnotationSet}
\alias{setAttributesBySampleAnnotationSet.AromaMicroarrayDataSet}
\alias{setAttributesByTags.AromaMicroarrayDataFile}
\alias{setAttributesByTags.AromaTabularBinaryFile}
\alias{setAttributeXY}
\alias{setAttributeXY.AromaMicroarrayDataFile}
\alias{setCytoband}
\alias{setCytoband.ChromosomeExplorer}
\alias{setGenome}
\alias{setGenome.ChromosomalModel}
\alias{setLocusFields}
\alias{setLocusFields.RawGenomicSignals}
\alias{setName.RawGenomicSignals}
\alias{setParallelSafe}
\alias{setParallelSafe.Explorer}
\alias{setRam}
\alias{setRam.AromaSettings}
\alias{setReportPathPattern}
\alias{setReportPathPattern.Explorer}
\alias{setSigma}
\alias{setSigma.RawGenomicSignals}
\alias{setStateColorMap}
\alias{setStateColorMap.SegmentedGenomicSignalsInterface}
\alias{setStates}
\alias{setStates.SegmentedGenomicSignalsInterface}
\alias{setSubname}
\alias{setSubname.Explorer}
\alias{setTags.AromaTransform}
\alias{setup}
\alias{setup.ChromosomeExplorer}
\alias{setup.Explorer}
\alias{setVerbose}
\alias{setVerbose.AromaSettings}
\alias{setWeights}
\alias{setWeights.RawGenomicSignals}
\alias{setXScale}
\alias{setXScale.RawGenomicSignals}
\alias{setYScale}
\alias{setYScale.RawGenomicSignals}
\alias{setZooms}
\alias{setZooms.ChromosomeExplorer}
\alias{shakyText}
\alias{shakyText.default}
\alias{signalRange}
\alias{signalRange.RawGenomicSignals}
\alias{smoothWRMA}
\alias{smoothWRMA.matrix}
\alias{smoothWSA}
\alias{smoothWSA.matrix}
\alias{sort.RawGenomicSignals}
\alias{splitByReportPathPattern}
\alias{splitByReportPathPattern.Explorer}
\alias{splitInChunks}
\alias{splitInChunks.numeric}
\alias{sqrtabs}
\alias{sqrtcenter}
\alias{sqrtneg}
\alias{sqrtpos}
\alias{stextChipType}
\alias{stextChipType.character}
\alias{stextLabel}
\alias{stextLabel.AromaMicroarrayDataFile}
\alias{stextLabels}
\alias{stextLabels.AromaMicroarrayDataFile}
\alias{stextSize}
\alias{stextSize.AromaMicroarrayDataFile}
\alias{stringTree}
\alias{stringTree.character}
\alias{subsample}
\alias{subsample.BinnedScatter}
\alias{subset.AromaTabularBinaryFile}
\alias{subset.BinnedScatter}
\alias{subtractBy}
\alias{subtractBy.RawGenomicSignals}
\alias{summary.AromaTabularBinaryFile}
\alias{summary.RawGenomicSignals}
\alias{summaryOfUnits}
\alias{summaryOfUnits.AromaUflFile}
\alias{text.ProbePositionEffects}
\alias{textSequence}
\alias{textSequence.ProbePositionEffects}
\alias{TextUnitNamesFile}
\alias{updateData}
\alias{updateData.AromaTabularBinaryFile}
\alias{updateDataColumn}
\alias{updateDataColumn.AromaTabularBinaryFile}
\alias{updateGenotypes}
\alias{updateGenotypes.AromaUnitGenotypeCallFile}
\alias{updatePositions}
\alias{updatePositions.AromaCellPositionFile}
\alias{updateSequenceMatrix}
\alias{updateSequenceMatrix.AromaCellSequenceFile}
\alias{updateSequences}
\alias{updateSequences.AromaCellSequenceFile}
\alias{updateTargetStrands}
\alias{updateTargetStrands.AromaCellSequenceFile}
\alias{validate.AromaMicroarrayDataSet}
\alias{validate.AromaUnitSignalBinarySet}
\alias{weightedMean}
\alias{whatDataType}
\alias{whatDataType.default}
\alias{writeAxesLayers}
\alias{writeAxesLayers.ChromosomeExplorer}
\alias{writeCopyNumberRegionLayers}
\alias{writeCopyNumberRegionLayers.ChromosomeExplorer}
\alias{writeCytobandLayers}
\alias{writeCytobandLayers.ChromosomeExplorer}
\alias{writeFooter}
\alias{writeGraphs}
\alias{writeGraphs.ChromosomeExplorer}
\alias{writeGridHorizontalLayers}
\alias{writeGridHorizontalLayers.ChromosomeExplorer}
\alias{writeRawCopyNumberLayers}
\alias{writeRawCopyNumberLayers.ChromosomeExplorer}
\alias{writeRawFooter}
\alias{writeRawFooter.AromaTabularBinaryFile}
\alias{writeRegions}
\alias{writeRegions.ChromosomeExplorer}
\alias{writeRegions.CopyNumberSegmentationModel}
\alias{writeRegions.GladModel}
\alias{xMax}
\alias{xMax.RawGenomicSignals}
\alias{xMin}
\alias{xMin.RawGenomicSignals}
\alias{xmlToList}
\alias{xmlToList.character}
\alias{xRange}
\alias{xRange.RawGenomicSignals}
\alias{xSeq}
\alias{xSeq.RawGenomicSignals}
\alias{yellow.colors}
\alias{yMax}
\alias{yMax.RawGenomicSignals}
\alias{yMin}
\alias{yMin.RawGenomicSignals}
\alias{yRange}
\alias{yRange.RawGenomicSignals}
\alias{getDefaultFullName.AromaMicroarrayDataSet}
\alias{getDefaultFullName.AromaTabularBinarySet}
\alias{drawCnRegions.MPCBS}
\alias{extractCopyNumberRegions.MPCBS}
\alias{segmentByMPCBS}
\alias{getAromaUgpFile.AromaPlatformInterface}
\alias{isCompatibleWith}
\alias{isCompatibleWith.AromaPlatformInterface}
\alias{getTags.Arguments}
\alias{doCBS}
\alias{doCBS.CopyNumberDataSet}
\alias{doCBS.CopyNumberDataSetTuple}
\alias{getOutputSet}
\alias{getOutputSet.ChromosomalModel}
\alias{setAttributesBy.AromaTabularBinarySet}
\alias{setAttributesBySampleAnnotationFile.AromaTabularBinarySet}
\alias{setAttributesBySampleAnnotationSet.AromaTabularBinarySet}
\alias{setAttributesByTags.AromaTabularBinarySet}
\alias{allocate.AromaCellCpgFile}
\alias{byChipType.AromaCellCpgFile}
\alias{getColumnNames.AromaCellCpgFile}
\alias{getFilenameExtension.AromaCellCpgFile}
\alias{isMissing.AromaCellCpgFile}
\alias{readCpgs}
\alias{readCpgs.AromaCellCpgFile}
\alias{updateCpgs}
\alias{updateCpgs.AromaCellCpgFile}
\alias{getRandomSeed}
\alias{getRandomSeed.CbsModel}
\alias{setRandomSeed}
\alias{setRandomSeed.CbsModel}
\alias{equals.CopyNumberRegions}
\alias{writeDataFrame}
\alias{getParentName}
\alias{getParentName.GenericDataFile}
\alias{getParentName.GenericDataFileSet}
\alias{getRootName}
\alias{getRootName.AromaTabularBinaryFile}
\alias{getRootName.AromaTabularBinarySet}
\alias{getAromaFullNameTranslatorSet}
\alias{getAromaFullNameTranslatorSet.character}
\alias{getFullNameTranslatorSet}
\alias{getFullNameTranslatorSet.character}
\alias{getFullNameTranslatorSet.GenericDataFileSet}
\alias{getAromaFullNameTranslatorSet.AromaMicroarrayDataSet}
\alias{getAromaFullNameTranslatorSet.AromaUnitSignalBinarySet}
\alias{extractRegions}
\alias{extractRegions.RawGenomicSignals}
\alias{subset.CopyNumberRegions}
\alias{writeDataFrame.AromaUnitTotalCnBinarySet}
\alias{writeDataFrame.AromaUnitFracBCnBinarySet}
\alias{-.CopyNumberRegions}
\alias{*.CopyNumberRegions}
\alias{+.CopyNumberRegions}
\alias{getMeanAt}
\alias{getMeanAt.CopyNumberRegions}
\alias{getValueAt}
\alias{getValueAt.CopyNumberRegions}
\alias{prune}
\alias{prune.CopyNumberRegions}
\alias{simulateRawCopyNumbers}
\alias{simulateRawCopyNumbers.CopyNumberRegions}
\alias{xRange.CopyNumberRegions}
\alias{xMax.CopyNumberRegions}
\alias{xMin.CopyNumberRegions}
\alias{append.CopyNumberRegions}
\alias{as.character.SegmentationDataSet}
\alias{byName.CbsSegmentationDataSet}
\alias{byPath.CbsSegmentationDataSet}
\alias{byPath.SegmentationDataSet}
\alias{CbsSegmentationDataFile}
\alias{CbsSegmentationDataSet}
\alias{extractByReferenceName}
\alias{extractByReferenceName.SegmentationDataSet}
\alias{extractBySampleName}
\alias{extractBySampleName.SegmentationDataSet}
\alias{extractCopyNumberRegions.SegmentationDataFile}
\alias{extractCopyNumberRegions.SegmentationDataSet}
\alias{getChipType.SegmentationDataSet}
\alias{getChromosome.SegmentationDataFile}
\alias{getChromosomes.CopyNumberRegions}
\alias{getChromosomes.SegmentationDataSet}
\alias{getDefaultFullName.SegmentationDataSet}
\alias{getReferenceName}
\alias{getReferenceName.SegmentationDataFile}
\alias{getReferenceNames}
\alias{getReferenceNames.SegmentationDataSet}
\alias{getSampleName}
\alias{getSampleName.SegmentationDataFile}
\alias{getSampleNames}
\alias{getSampleNames.SegmentationDataSet}
\alias{loadFit}
\alias{loadFit.CbsSegmentationDataFile}
\alias{loadFit.SegmentationDataFile}
\alias{nbrOfChromosomes}
\alias{nbrOfChromosomes.CopyNumberRegions}
\alias{SegmentationDataFile}
\alias{SegmentationDataSet}
\alias{unique.CopyNumberRegions}
\alias{byGenome}
\alias{byGenome.AromaGenomeTextFile}
\alias{findByGenome}
\alias{findByGenome.AromaGenomeTextFile}
\alias{readDataFrame.AromaGenomeTextFile}
\alias{-.RawGenomicSignals}
\alias{*.RawGenomicSignals}
\alias{+.RawGenomicSignals}
\alias{drawDensity}
\alias{drawDensity.CopyNumberRegions}
\alias{drawDensity.RawGenomicSignals}
\alias{getDensity}
\alias{getDensity.CopyNumberRegions}
\alias{getLength}
\alias{getLength.CopyNumberRegions}
\alias{getOptionalArguments}
\alias{getOptionalArguments.CopyNumberChromosomalModel}
\alias{getChromosomeLength}
\alias{getChromosomeLength.CopyNumberChromosomalModel}
\alias{requireWithMemory}
\alias{requireWithMemory.default}
\alias{writeImage}
\alias{writeImage.Image}
\alias{createImage}
\alias{createImage.matrix}
\alias{getImageData}
\alias{getImageData.Image}
\alias{setImageData}
\alias{setImageData.Image}
\alias{RasterImage}
\alias{[.RasterImage}
\alias{as.character.RasterImage}
\alias{display.RasterImage}
\alias{dropAlpha}
\alias{dropAlpha.RasterImage}
\alias{getImageData.RasterImage}
\alias{read}
\alias{read.RasterImage}
\alias{rescale.RasterImage}
\alias{save.RasterImage}
\alias{setImageData.RasterImage}
\alias{str.RasterImage}
\alias{write.RasterImage}
\alias{writeImage.RasterImage}
\alias{colorize.RasterImage}
\alias{interleave.RasterImage}
\alias{dropRootPathTags.default}
\alias{loadAll}
\alias{loadAll.SampleAnnotationSet}
\alias{byPathnames}
\alias{byPathnames.SampleAnnotationSet}
\alias{getAromaGenomeTextFile}
\alias{getAromaGenomeTextFile.ChromosomalModel}
\alias{getChromosomeLabels}
\alias{getChromosomeLabels.ChromosomeExplorer}
\alias{cat}
\alias{getOption}
\alias{append}
\alias{lapply}
\alias{sapply}
\alias{downloadACS}
\alias{downloadACS.AromaRepository}
\alias{downloadCDF}
\alias{downloadCDF.AromaRepository}
\alias{downloadChipTypeFile}
\alias{downloadProbeSeqsTXT}
\alias{downloadProbeSeqsTXT.AromaRepository}
\alias{downloadTXT}
\alias{downloadTXT.AromaRepository}
\alias{downloadUFL}
\alias{downloadUFL.AromaRepository}
\alias{downloadUGP}
\alias{downloadUGP.AromaRepository}
\alias{getUrlPath}
\alias{getUrlPath.AromaRepository}
\alias{getVerbose.AromaRepository}
\alias{setVerbose.AromaRepository}
\alias{downloadACC}
\alias{downloadACC.AromaRepository}
\alias{downloadACM}
\alias{downloadACM.AromaRepository}
\alias{downloadACP}
\alias{downloadACP.AromaRepository}
\alias{extractPSCNArray}
\alias{extractPSCNArray.AromaUnitTotalCnBinaryFile}
\alias{extractPSCNArray.AromaUnitTotalCnBinarySet}
\alias{extractPSCNMatrix}
\alias{extractPSCNMatrix.AromaUnitTotalCnBinaryFile}
\alias{AromaUcscGenomeTextFile}
\alias{findByGenome.AromaUcscGenomeTextFile}
\alias{getDefaultExtension}
\alias{getDefaultExtension.UnitAnnotationDataFile}
\alias{readDataFrame.AromaUcscGenomeTextFile}
\alias{getDefaultExtension.AromaCellCpgFile}
\alias{getDefaultExtension.AromaCellPositionFile}
\alias{getDefaultExtension.AromaCellSequenceFile}
\alias{getDefaultExtension.AromaUflFile}
\alias{getDefaultExtension.AromaUgpFile}
\alias{plotCoverageMap}
\alias{plotCoverageMap.AromaUgpFile}
\alias{assertOneChromosome}
\alias{assertOneChromosome.RawGenomicSignals}
\alias{getCXY}
\alias{getCXY.RawGenomicSignals}
\alias{getChromosomes.RawGenomicSignals}
\alias{nbrOfChromosomes.RawGenomicSignals}
\alias{extractChromosome}
\alias{extractChromosome.RawGenomicSignals}
\alias{extractChromosomes}
\alias{extractChromosomes.RawGenomicSignals}
\alias{clearCache.RawGenomicSignals}
\alias{clone.RawGenomicSignals}
\alias{getBasicField}
\alias{getBasicField.RawGenomicSignals}
\alias{print.RawGenomicSignals}
\alias{setBasicField}
\alias{setBasicField.RawGenomicSignals}
\alias{newInstance.RawGenomicSignals}
\alias{getDefaultLocusFields}
\alias{getDefaultLocusFields.RawGenomicSignals}
\alias{getVirtualLocusFields}
\alias{getVirtualLocusFields.RawGenomicSignals}
\alias{getVirtualLocusFields.SegmentedGenomicSignalsInterface}
\alias{getVersion.Explorer}
\alias{updateSetupExplorerFile}
\alias{updateSetupExplorerFile.Explorer}
\alias{$.RichDataFrame}
\alias{$<-.RichDataFrame}
\alias{RichDataFrame}
\alias{[.RichDataFrame}
\alias{[[.RichDataFrame}
\alias{[[<-.RichDataFrame}
\alias{as.data.frame.RichDataFrame}
\alias{as.list.RichDataFrame}
\alias{binnedSmoothingByField}
\alias{binnedSmoothingByField.RawGenomicSignals}
\alias{dim.RichDataFrame}
\alias{dropVirtualColumn}
\alias{dropVirtualColumn.RichDataFrame}
\alias{getColumnNames.RichDataFrame}
\alias{getColumnNamesTranslator}
\alias{getColumnNamesTranslator.RichDataFrame}
\alias{getFullName.RichDataFrame}
\alias{getGenericSummary}
\alias{getGenericSummary.RichDataFrame}
\alias{getName.RichDataFrame}
\alias{getSignalColumnName}
\alias{getSignalColumnName.RawGenomicSignals}
\alias{getSignalColumnNames}
\alias{getSignalColumnNames.RawGenomicSignals}
\alias{getSignals.RawCopyNumbers}
\alias{getTags.RichDataFrame}
\alias{getVirtualColumn}
\alias{getVirtualColumn.RichDataFrame}
\alias{getVirtualColumnFunction}
\alias{getVirtualColumnFunction.RichDataFrame}
\alias{getVirtualColumnFunctions}
\alias{getVirtualColumnFunctions.RichDataFrame}
\alias{getVirtualColumnNames}
\alias{getVirtualColumnNames.RichDataFrame}
\alias{getVirtualField}
\alias{getVirtualField.SegmentedGenomicSignalsInterface}
\alias{hasColumn}
\alias{hasColumn.RichDataFrame}
\alias{hasColumns}
\alias{hasColumns.RichDataFrame}
\alias{hasVirtualColumn}
\alias{hasVirtualColumn.RichDataFrame}
\alias{hasVirtualColumns}
\alias{hasVirtualColumns.RichDataFrame}
\alias{length.RichDataFrame}
\alias{names.RichDataFrame}
\alias{newInstance.RichDataFrame}
\alias{print.RichDataFrame}
\alias{rbind.RichDataFrame}
\alias{setAttributes.RichDataFrame}
\alias{setColumnNamesMap}
\alias{setColumnNamesMap.RichDataFrame}
\alias{setColumnNamesTranslator}
\alias{setColumnNamesTranslator.RichDataFrame}
\alias{setName.RichDataFrame}
\alias{setSignals}
\alias{setSignals.RawGenomicSignals}
\alias{setTags.RichDataFrame}
\alias{setVirtualColumn}
\alias{setVirtualColumn.RichDataFrame}
\alias{setVirtualColumnFunctions}
\alias{setVirtualColumnFunctions.RichDataFrame}
\alias{subset.RichDataFrame}
\alias{translateColumnNames.RichDataFrame}
\alias{as.NonPairedPSCNData}
\alias{as.NonPairedPSCNData.NonPairedPSCNData}
\alias{as.NonPairedPSCNData.data.frame}
\alias{as.PairedPSCNData}
\alias{as.PairedPSCNData.NonPairedPSCNData}
\alias{as.PairedPSCNData.PairedPSCNData}
\alias{as.PairedPSCNData.data.frame}
\alias{callNaiveGenotypes.NonPairedPSCNData}
\alias{callNaiveGenotypes.PairedPSCNData}
\alias{callSNPs}
\alias{callSNPs.AbstractPSCNData}
\alias{callSegmentationOutliers}
\alias{callSegmentationOutliers.NonPairedPSCNData}
\alias{callSegmentationOutliers.PairedPSCNData}
\alias{dropSegmentationOutliers}
\alias{dropSegmentationOutliers.NonPairedPSCNData}
\alias{dropSegmentationOutliers.PairedPSCNData}
\alias{extractNonPairedPSCNData}
\alias{extractNonPairedPSCNData.PairedPSCNData}
\alias{findLargeGaps}
\alias{findLargeGaps.AbstractCNData}
\alias{getChipType.AbstractCNData}
\alias{getLocusData}
\alias{getLocusData.AbstractCNData}
\alias{getPlatform.AbstractCNData}
\alias{getSNPFields}
\alias{getSNPFields.AbstractPSCNData}
\alias{getSignalColumnNames.NonPairedPSCNData}
\alias{getSignalColumnNames.PairedPSCNData}
\alias{getTCNs}
\alias{getTCNs.PairedPSCNData}
\alias{getTotalCopyNumbers}
\alias{getTotalCopyNumbers.PairedPSCNData}
\alias{hasKnownPositions}
\alias{hasKnownPositions.AbstractCNData}
\alias{normalizeTumorBoost.PairedPSCNData}
\alias{orderAlongGenome}
\alias{orderAlongGenome.AbstractCNData}
\alias{segmentByCBS.NonPairedPSCNData}
\alias{segmentByCBS.PairedPSCNData}
\alias{segmentByPairedPSCBS}
\alias{segmentByPairedPSCBS.PairedPSCNData}
\alias{setChipType}
\alias{setChipType.AbstractCNData}
\alias{setPlatform}
\alias{setPlatform.AbstractCNData}
\alias{normalizeTumorBoost}
\alias{callNaiveGenotypes}
\alias{getNumberOfFilesAveraged.AromaUnitSignalBinaryFile}
\alias{allocate.AromaUnitPscnBinaryFile}
\alias{byName.AromaUnitPscnBinarySet}
\alias{extractRawCopyNumbers.AromaUnitPscnBinaryFile}
\alias{getAverageFile.AromaUnitPscnBinarySet}
\alias{getColumnNames.AromaUnitPscnBinaryFile}
\alias{hasAlleleBFractions.AromaUnitPscnBinaryFile}
\alias{hasStrandiness.AromaUnitPscnBinaryFile}
\alias{hasTotalCopyNumberRatios}
\alias{hasTotalCopyNumberRatios.AromaUnitPscnBinaryFile}
\alias{writeDataFrame.AromaUnitPscnBinarySet}
\alias{exportAromaUnitPscnBinarySet}
\alias{exportAromaUnitPscnBinarySet.AromaUnitTotalCnBinarySet}
\alias{exportAromaUnitPscnBinarySet.list}
\alias{downloadAll}
\alias{downloadAll.AromaRepository}
\alias{listFiles}
\alias{clearCache.AromaRepository}
\alias{findAnnotationDataByChipType.AromaRepository}
\alias{downloadUGC}
\alias{downloadUGC.AromaRepository}
\alias{AromaUnitGcContentFile}
\alias{getColumnNames.AromaUnitGcContentFile}
\alias{getExtensionPattern.AromaUnitGcContentFile}
\alias{getFilenameExtension.AromaUnitGcContentFile}
\alias{AromaUgcFile}
\alias{getReference}
\alias{getReference.CopyNumberChromosomalModel}
\alias{setReference}
\alias{setReference.CopyNumberChromosomalModel}
\alias{getMaxNAFraction}
\alias{getMaxNAFraction.CopyNumberChromosomalModel}
\alias{getDefaultColumnNames.AromaCellCpgFile}
\alias{getDefaultColumnNames.AromaCellPositionFile}
\alias{getDefaultColumnNames.AromaCellSequenceFile}
\alias{getDefaultColumnNames.AromaTabularBinaryFile}
\alias{getDefaultColumnNames.AromaUflFile}
\alias{getDefaultColumnNames.AromaUgpFile}
\alias{getDefaultColumnNames.AromaUnitChromosomeTabularBinaryFile}
\alias{getDefaultColumnNames.AromaUnitGcContentFile}
\alias{getDefaultColumnNames.AromaUnitPscnBinaryFile}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
