%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParametersInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{ParametersInterface}
\docType{class}
\alias{ParametersInterface}


 \title{The ParametersInterface class interface}

 \description{
  Package:  aroma.core \cr
\bold{Class ParametersInterface}\cr

\code{\link[R.oo]{Interface}}\cr
\code{~~|}\cr
\code{~~+--}\code{ParametersInterface}\cr

\bold{Directly known subclasses:}\cr
\emph{\link[aroma.core]{AromaTransform}}\cr

public class \bold{ParametersInterface}\cr
extends \link[R.oo]{Interface}\cr


 }

 \usage{
ParametersInterface(...)
}

 \arguments{
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{getParameters} \tab  -\cr
 \tab \code{getParametersAsString} \tab  -\cr
}


 \bold{Methods inherited from Interface}:\cr
extend, print, uses


 }

 \author{Henrik Bengtsson}
\keyword{classes}
