%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Explorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.Explorer}
\alias{process.Explorer}
\alias{Explorer.process}
\alias{process.Explorer}
\alias{process,Explorer-method}


\title{Generates image files, scripts and dynamic pages for the explorer}

\description{
 Generates image files, scripts and dynamic pages for the explorer.
}

\usage{\method{process}{Explorer}(...)}

\arguments{
  \item{arrays}{A \code{\link[base]{vector}} of arrays specifying which arrays to
   be considered.  If \code{\link[base]{NULL}}, all are processed.}
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Explorer}}.
}
\keyword{internal}
\keyword{methods}
